+++ services/system/AccountSys/ui/public/loader/assets/index-IMu5n1Ph.js
@@ -1,0 +1,19 @@
+(function(){const r=document.createElement("link").relList;if(r&&r.supports&&r.supports("modulepreload"))return;for(const n of document.querySelectorAll('link[rel="modulepreload"]'))i(n);new MutationObserver(n=>{for(const t of n)if(t.type==="childList")for(const s of t.addedNodes)s.tagName==="LINK"&&s.rel==="modulepreload"&&i(s)}).observe(document,{childList:!0,subtree:!0});function o(n){const t={};return n.integrity&&(t.integrity=n.integrity),n.referrerPolicy&&(t.referrerPolicy=n.referrerPolicy),n.crossOrigin==="use-credentials"?t.credentials="include":n.crossOrigin==="anonymous"?t.credentials="omit":t.credentials="same-origin",t}function i(n){if(n.ep)return;n.ep=!0;const t=o(n);fetch(n.href,t)}})();const F="modulepreload",I=function(e){return"/loader/"+e},b={},O=function(r,o,i){let n=Promise.resolve();if(o&&o.length>0){const t=document.getElementsByTagName("link");n=Promise.all(o.map(s=>{if(s=I(s),s in b)return;b[s]=!0;const a=s.endsWith(".css"),g=a?'[rel="stylesheet"]':"";if(!!i)for(let f=t.length-1;f>=0;f--){const l=t[f];if(l.href===s&&(!a||l.rel==="stylesheet"))return}else if(document.querySelector(`link[href="${s}"]${g}`))return;const c=document.createElement("link");if(c.rel=a?"stylesheet":F,a||(c.as="script",c.crossOrigin=""),c.href=s,document.head.appendChild(c),a)return new Promise((f,l)=>{c.addEventListener("load",f),c.addEventListener("error",()=>l(new Error(`Unable to preload CSS for ${s}`)))})}))}return n.then(()=>r()).catch(t=>{const s=new Event("vite:preloadError",{cancelable:!0});if(s.payload=t,window.dispatchEvent(s),!s.defaultPrevented)throw t})},_=`function derp(num) {
+  return num * 2;
+}
+
+export function prnt(string) {
+  console.log(\`from imported code: john \${derp(2)}\`, string);
+}
+`;var y;(function(e){e.Call="call",e.Reply="reply",e.Syn="syn",e.SynAck="synAck",e.Ack="ack"})(y||(y={}));var v;(function(e){e.Fulfilled="fulfilled",e.Rejected="rejected"})(v||(v={}));var R;(function(e){e.ConnectionDestroyed="ConnectionDestroyed",e.ConnectionTimeout="ConnectionTimeout",e.NoIframeSrc="NoIframeSrc"})(R||(R={}));var A;(function(e){e.DataCloneError="DataCloneError"})(A||(A={}));var w;(function(e){e.Message="message"})(w||(w={}));const j=(e,r)=>{const o=[];let i=!1;return{destroy(n){i||(i=!0,r(`${e}: Destroying connection`),o.forEach(t=>{t(n)}))},onDestroy(n){i?n():o.push(n)}}},x=e=>(...r)=>{e&&console.log("[Penpal]",...r)},N=({name:e,message:r,stack:o})=>({name:e,message:r,stack:o}),K=e=>{const r=new Error;return Object.keys(e).forEach(o=>r[o]=e[o]),r},V=(e,r,o)=>{const{localName:i,local:n,remote:t,originForSending:s,originForReceiving:a}=e;let g=!1;const h=c=>{if(c.source!==t||c.data.penpal!==y.Call)return;if(a!=="*"&&c.origin!==a){o(`${i} received message from origin ${c.origin} which did not match expected origin ${a}`);return}const f=c.data,{methodName:l,args:d,id:p}=f;o(`${i}: Received ${l}() call`);const C=u=>m=>{if(o(`${i}: Sending ${l}() reply`),g){o(`${i}: Unable to send ${l}() reply due to destroyed connection`);return}const E={penpal:y.Reply,id:p,resolution:u,returnValue:m};u===v.Rejected&&m instanceof Error&&(E.returnValue=N(m),E.returnValueIsError=!0);try{t.postMessage(E,s)}catch(M){if(M.name===A.DataCloneError){const $={penpal:y.Reply,id:p,resolution:v.Rejected,returnValue:N(M),returnValueIsError:!0};t.postMessage($,s)}throw M}};new Promise(u=>u(r[l].apply(r,d))).then(C(v.Fulfilled),C(v.Rejected))};return n.addEventListener(w.Message,h),()=>{g=!0,n.removeEventListener(w.Message,h)}};let H=0;const B=()=>++H,T=".",k=e=>e?e.split(T):[],z=e=>e.join(T),U=(e,r)=>{const o=k(r||"");return o.push(e),z(o)},Y=(e,r,o)=>{const i=k(r);return i.reduce((n,t,s)=>(typeof n[t]>"u"&&(n[t]={}),s===i.length-1&&(n[t]=o),n[t]),e),e},D=(e,r)=>{const o={};return Object.keys(e).forEach(i=>{const n=e[i],t=U(i,r);typeof n=="object"&&Object.assign(o,D(n,t)),typeof n=="function"&&(o[t]=n)}),o},q=e=>{const r={};for(const o in e)Y(r,o,e[o]);return r},W=(e,r,o,i,n)=>{const{localName:t,local:s,remote:a,originForSending:g,originForReceiving:h}=r;let c=!1;n(`${t}: Connecting call sender`);const f=d=>(...p)=>{n(`${t}: Sending ${d}() call`);let C;try{a.closed&&(C=!0)}catch{C=!0}if(C&&i(),c){const u=new Error(`Unable to send ${d}() call due to destroyed connection`);throw u.code=R.ConnectionDestroyed,u}return new Promise((u,m)=>{const E=B(),M=S=>{if(S.source!==a||S.data.penpal!==y.Reply||S.data.id!==E)return;if(h!=="*"&&S.origin!==h){n(`${t} received message from origin ${S.origin} which did not match expected origin ${h}`);return}const P=S.data;n(`${t}: Received ${d}() reply`),s.removeEventListener(w.Message,M);let L=P.returnValue;P.returnValueIsError&&(L=K(L)),(P.resolution===v.Fulfilled?u:m)(L)};s.addEventListener(w.Message,M);const $={penpal:y.Call,id:E,methodName:d,args:p};a.postMessage($,g)})},l=o.reduce((d,p)=>(d[p]=f(p),d),{});return Object.assign(e,q(l)),()=>{c=!0}},Q=(e,r)=>{let o;return e!==void 0&&(o=window.setTimeout(()=>{const i=new Error(`Connection timed out after ${e}ms`);i.code=R.ConnectionTimeout,r(i)},e)),()=>{clearTimeout(o)}},G=(e,r,o,i)=>{const{destroy:n,onDestroy:t}=o;return s=>{if(!(e instanceof RegExp?e.test(s.origin):e==="*"||e===s.origin)){i(`Child: Handshake - Received SYN-ACK from origin ${s.origin} which did not match expected origin ${e}`);return}i("Child: Handshake - Received SYN-ACK, responding with ACK");const g=s.origin==="null"?"*":s.origin,h={penpal:y.Ack,methodNames:Object.keys(r)};window.parent.postMessage(h,g);const c={localName:"Child",local:window,remote:window.parent,originForSending:g,originForReceiving:s.origin},f=V(c,r,i);t(f);const l={},d=W(l,c,s.data.methodNames,n,i);return t(d),l}},J=()=>{try{clearTimeout()}catch{return!1}return!0},X=(e={})=>{const{parentOrigin:r="*",methods:o={},timeout:i,debug:n=!1}=e,t=x(n),s=j("Child",t),{destroy:a,onDestroy:g}=s,h=D(o),c=G(r,h,s,t),f=()=>{t("Child: Handshake - Sending SYN");const d={penpal:y.Syn},p=r instanceof RegExp?"*":r;window.parent.postMessage(d,p)};return{promise:new Promise((d,p)=>{const C=Q(i,a),u=m=>{if(J()&&!(m.source!==parent||!m.data)&&m.data.penpal===y.SynAck){const E=c(m);E&&(window.removeEventListener(w.Message,u),C(),d(E))}};window.addEventListener(w.Message,u),f(),g(m=>{window.removeEventListener(w.Message,u),m&&p(m)})}),destroy(){a()}}};document.querySelector("#app").innerHTML=`
+  <div>
+    <h1>App 2 / Loader</h1>
+    <p>This is a generated SPA designed to act as the loader, its sole purpose is to be rendered in an iframe and run a WASM Component, then execute functions in the WASM component and send the results back to its parent iframe.</p>
+  </div>
+`;const Z=e=>typeof e=="object"&&e!==null&&"service"in e&&"method"in e&&"params"in e,ee=async e=>{if(!Z(e))throw new Error("Invalid function call param.");const{load:r}=await O(()=>import("./index-O1hlgyW2.js"),__vite__mapDeps([])),i=await fetch("/loader/functions.wasm").then(s=>s.arrayBuffer()),t=await r(i,[{"component:account-sys/imports":_}]);return console.log(t,"is the module I have"),t[e.method](...e.params)},ne=X({methods:{functionCall:ee}});ne.promise.then(e=>{e.add(3,1).then(r=>console.log(r))});export{O as _};
+function __vite__mapDeps(indexes) {
+  if (!__vite__mapDeps.viteFileDeps) {
+    __vite__mapDeps.viteFileDeps = []
+  }
+  return indexes.map((i) => __vite__mapDeps.viteFileDeps[i])
+}
