+++ services/system/AccountSys/ui/public/loader/assets/index-O1hlgyW2.js
@@ -1,4 +1,4 @@
-var no=(n,e,t)=>{if(!e.has(n))throw TypeError("Cannot "+t)};var xe=(n,e,t)=>(no(n,e,"read from private field"),t?t.call(n):e.get(n)),as=(n,e,t)=>{if(e.has(n))throw TypeError("Cannot add the same private member more than once");e instanceof WeakSet?e.add(n):e.set(n,t)},vi=(n,e,t,s)=>(no(n,e,"write to private field"),s?s.call(n,t):e.set(n,t),t);import{_ as Ra}from"./index-pZAdW6pa.js";function Ic(n){return new Promise(e=>{setTimeout(e,n)})}let so="https://unpkg.com";const ls=new Map;async function Qi(n){if(ls.has(n))return ls.get(n);await Ic(200);const e=fetch(n).then(async t=>{if(t.ok)return{url:t.url,body:await t.text()};throw new Error(await t.text())}).catch(t=>{throw ls.delete(n),t});return ls.set(n,e),e}async function xi(n){return(await Qi(n)).url}const $c=new TextDecoder,Sc=n=>({name:"loader",async resolveId(e,t){if(n.find(s=>s[0]==e)||(e.endsWith("/")&&(e=e.slice(0,-1)),e.startsWith("http:")||e.startsWith("https:")))return e;if(e.startsWith(".")){const s=new URL(e,t).href;return await xi(s)}else{try{const s=await xi(`${so}/${e}/package.json`),i=(await Qi(s)).body,r=JSON.parse(i);if(r.svelte||r.module||r.main){const o=s.replace(/\/package\.json$/,"");return new URL(r.svelte||r.module||r.main,`${o}/`).href}}catch{}return await xi(`${so}/${e}`)}},async load(e){let t=n.find(i=>i[0]==e);return t?typeof t[1]=="string"?t[1]:$c.decode(t[1]):(await Qi(e)).body},async transform(e,t){if(!t.endsWith(".js"))return;let s=e.match(/new URL\((.*)/g);if(!(!s||!s.length))return s.forEach(i=>{let r=i.match(/'.\/(.*)'/)[1];if(!r.endsWith(".wasm"))return;let o=n.find(l=>l[0]==r);if(!o)return;let a=URL.createObjectURL(new Blob([o[1]],{type:"application/wasm"}));!a||a=="undefined"||(e=e.replace(`new URL('./${r}', import.meta.url)`,`'${a}'`))}),{code:e,map:{mappings:""}}}});let La=0;const Oa=Symbol.dispose||Symbol.for("dispose"),Nc=class{constructor(e){this.msg=e}toDebugString(){return this.msg}};let Pc=class{constructor(e){e||console.trace("no handler"),this.id=++La,this.handler=e}read(e){return this.handler.read?this.handler.read(e):this.handler.blockingRead.call(this,e)}blockingRead(e){return this.handler.blockingRead.call(this,e)}skip(e){if(this.handler.skip)return this.handler.skip.call(this,e);if(this.handler.read){const t=this.handler.read.call(this,e);return BigInt(t.byteLength)}return this.blockingSkip.call(this,e)}blockingSkip(e){if(this.handler.blockingSkip)return this.handler.blockingSkip.call(this,e);const t=this.handler.blockingRead.call(this,e);return BigInt(t.byteLength)}subscribe(){console.log(`[streams] Subscribe to input stream ${this.id}`)}[Oa](){this.handler.drop&&this.handler.drop.call(this)}},Cc=class{constructor(e){e||console.trace("no handler"),this.id=++La,this.open=!0,this.handler=e}checkWrite(e){return this.open?this.handler.checkWrite?this.handler.checkWrite.call(this,e):1000000n:0n}write(e){this.handler.write.call(this,e)}blockingWriteAndFlush(e){this.handler.write.call(this,e)}flush(){this.handler.flush&&this.handler.flush.call(this)}blockingFlush(){this.open=!0}writeZeroes(e){this.write.call(this,new Uint8Array(Number(e)))}blockingWriteZeroes(e){this.blockingWrite.call(this,new Uint8Array(Number(e)))}blockingWriteZeroesAndFlush(e){this.blockingWriteAndFlush.call(this,new Uint8Array(Number(e)))}splice(e,t){const s=Math.min(t,this.checkWrite.call(this)),i=e.read(s);return this.write.call(this,i),i.byteLength}blockingSplice(e,t){console.log(`[streams] Blocking splice ${this.id}`)}forward(e){console.log(`[streams] Forward ${this.id}`)}subscribe(){console.log(`[streams] Subscribe to output stream ${this.id}`)}[Oa](){}};const Mc={Error:Nc},Ir={InputStream:Pc,OutputStream:Cc},{InputStream:Rc,OutputStream:Lc}=Ir;let Oc=null,Dc={dir:{}};const Wt={seconds:BigInt(0),nanoseconds:0};function ki(n,e,t){e==="."&&io&&Tc(io[0])===n&&(e=Oc,e.startsWith("/")&&e!=="/"&&(e=e.slice(1)));let s=n,i;do{if(!s||!s.dir)throw"not-directory";i=e.indexOf("/");const r=i===-1?e:e.slice(0,i);if(r==="."||r==="")return s;if(r==="..")throw"no-entry";!s.dir[r]&&t.create?s=s.dir[r]=t.directory?{dir:{}}:{source:new Uint8Array([])}:s=s.dir[r]}while(i!==-1);if(!s)throw"no-entry";return s}function cs(n){return typeof n.source=="string"&&(n.source=new TextEncoder().encode(n.source)),n.source}let Da=class{constructor(e){this.idx=0,this.entries=e}readDirectoryEntry(){if(this.idx===this.entries.length)return null;const[e,t]=this.entries[this.idx];return this.idx+=1,{name:e,type:t.dir?"directory":"regular-file"}}};var qn,ke,Xn,Zn;let Zs=(Zn=class{constructor(e,t){as(this,qn,void 0);as(this,ke,void 0);as(this,Xn,0);t?vi(this,qn,e):vi(this,ke,e)}_getEntry(e){return xe(e,ke)}readViaStream(e){const t=cs(xe(this,ke));let s=Number(e);return new Rc({blockingRead(i){if(s===t.byteLength)throw{tag:"closed"};const r=t.slice(s,s+Number(i));return s+=r.byteLength,r}})}writeViaStream(e){const t=xe(this,ke);let s=Number(e);return new Lc({write(i){const r=new Uint8Array(i.byteLength+t.source.byteLength);return r.set(t.source,0),r.set(i,s),s+=i.byteLength,t.source=r,i.byteLength}})}appendViaStream(){console.log("[filesystem] APPEND STREAM")}advise(e,t,s,i){console.log("[filesystem] ADVISE",e,t,s,i)}syncData(){console.log("[filesystem] SYNC DATA")}getFlags(){console.log("[filesystem] FLAGS FOR")}getType(){return xe(this,qn)?"fifo":xe(this,ke).dir?"directory":xe(this,ke).source?"regular-file":"unknown"}setSize(e){console.log("[filesystem] SET SIZE",e)}setTimes(e,t){console.log("[filesystem] SET TIMES",e,t)}read(e,t){const s=cs(xe(this,ke));return[s.slice(t,t+e),t+e>=s.byteLength]}write(e,t){if(t!==0)throw"invalid-seek";return xe(this,ke).source=e,e.byteLength}readDirectory(){var e;if(!((e=xe(this,ke))!=null&&e.dir))throw"bad-descriptor";return new Da(Object.entries(xe(this,ke).dir).sort(([t],[s])=>t>s?1:-1))}sync(){console.log("[filesystem] SYNC")}createDirectoryAt(e){if(ki(xe(this,ke),e,{create:!0,directory:!0}).source)throw"exist"}stat(){let e="unknown",t=BigInt(0);if(xe(this,ke).source)e="directory";else if(xe(this,ke).dir){e="regular-file";const s=cs(xe(this,ke));t=BigInt(s.byteLength)}return{type:e,linkCount:BigInt(0),size:t,dataAccessTimestamp:Wt,dataModificationTimestamp:Wt,statusChangeTimestamp:Wt}}statAt(e,t){const s=ki(xe(this,ke),t);let i="unknown",r=BigInt(0);if(s.source){i="regular-file";const o=cs(s);r=BigInt(o.byteLength)}else s.dir&&(i="directory");return{type:i,linkCount:BigInt(0),size:r,dataAccessTimestamp:Wt,dataModificationTimestamp:Wt,statusChangeTimestamp:Wt}}setTimesAt(){console.log("[filesystem] SET TIMES AT")}linkAt(){console.log("[filesystem] LINK AT")}openAt(e,t,s,i,r){const o=ki(xe(this,ke),t,s);return new Zn(o)}readlinkAt(){console.log("[filesystem] READLINK AT")}removeDirectoryAt(){console.log("[filesystem] REMOVE DIR AT")}renameAt(){console.log("[filesystem] RENAME AT")}symlinkAt(){console.log("[filesystem] SYMLINK AT")}unlinkFileAt(){console.log("[filesystem] UNLINK FILE AT")}isSameObject(e){return e===this}metadataHash(){let e=BigInt(0);return e+=BigInt(xe(this,Xn)),{upper:e,lower:BigInt(0)}}metadataHashAt(e,t){let s=BigInt(0);return s+=BigInt(xe(this,Xn)),{upper:s,lower:BigInt(0)}}},qn=new WeakMap,ke=new WeakMap,Xn=new WeakMap,Zn);const Tc=Zs.prototype._getEntry;delete Zs.prototype._getEntry;let Ta=[[new Zs(Dc),"/"]],io=Ta[0];const Bc={getDirectories(){return Ta}},_c={Descriptor:Zs,DirectoryEntryStream:Da},{InputStream:Ba,OutputStream:Ys}=Ir,$r=Symbol.dispose??Symbol.for("dispose");let zc=[],Vc=[],Fc=null;const jc={getEnvironment(){return zc},getArguments(){return Vc},initialCwd(){return Fc}};class Uc extends Error{constructor(e){super(`Component exited ${e?"successfully":"with error"}`),this.exitError=!0,this.ok=e}}const Gc={exit(n){throw new Uc(n.tag==="err")}},Wc=new Ba({blockingRead(n){},subscribe(){},[$r](){}});let _a=new TextDecoder;const Hc=new Ys({write(n){console.log(_a.decode(n))},blockingFlush(){},[$r](){}}),qc=new Ys({write(n){console.error(_a.decode(n))},blockingFlush(){},[$r](){}}),Xc={InputStream:Ba,getStdin(){return Wc}},Zc={OutputStream:Ys,getStdout(){return Hc}},Yc={OutputStream:Ys,getStderr(){return qc}};let Sr=class{},Yn=class{};const Qc=new Yn,Kc=new Yn,Jc=new Sr,ed={TerminalInput:Sr},td={TerminalOutput:Yn},nd={TerminalOutput:Yn,getTerminalStderr(){return Kc}},sd={TerminalInput:Sr,getTerminalStdin(){return Jc}},id={TerminalOutput:Yn,getTerminalStdout(){return Qc}},Ii=65536;let $i,ro;const Ki={getRandomBytes(n){const e=new Uint8Array(Number(n));if(n>Ii)for(var t=0;t<n;t+=Ii)crypto.getRandomValues(e.slice(t,t+Ii));else crypto.getRandomValues(e);return e},getRandomU64(){return crypto.getRandomValues(new BigUint64Array(1))[0]},insecureRandom(){return $i===void 0&&($i=Ki.getRandomU64(),ro=Ki.getRandomU64()),[$i,ro]}},{getEnvironment:rd}=jc,{exit:od}=Gc,{getStderr:ad}=Yc,{getStdin:ld}=Xc,{getStdout:cd}=Zc,{TerminalInput:dd}=ed,{TerminalOutput:za}=td,{getTerminalStderr:ud}=nd,{getTerminalStdin:hd}=sd,{getTerminalStdout:pd}=id,{getDirectories:fd}=Bc,{Descriptor:ct,DirectoryEntryStream:Cf,filesystemErrorCode:md}=_c,{Error:un}=Mc,{InputStream:Qs,OutputStream:Nt}=Ir,{getRandomBytes:gd}=Ki,oo=n=>WebAssembly.compile(typeof Buffer<"u"?Buffer.from(n,"base64"):Uint8Array.from(atob(n),e=>e.charCodeAt(0)));class yd extends Error{constructor(e){const t=typeof e!="string";super(t?`${String(e)} (see error.payload)`:e),Object.defineProperty(this,"payload",{value:e,enumerable:t})}}let Si=new DataView(new ArrayBuffer);const k=n=>Si.buffer===n.buffer?Si:Si=new DataView(n.buffer),bd=typeof process<"u"&&process.versions&&process.versions.node;let Ni;async function ao(n){return bd?(Ni=Ni||await Ra(()=>import("./__vite-browser-external-XPXQPciS.js"),__vite__mapDeps([])),WebAssembly.compile(await Ni.readFile(n))):fetch(n).then(WebAssembly.compileStreaming)}function je(n){return n&&Ed.call(n,"payload")?n.payload:n}const Ed=Object.prototype.hasOwnProperty,ds=WebAssembly.instantiate,Me=Symbol.dispose||Symbol.for("dispose");function Ad(){throw new TypeError("Wasm uninitialized use `await $init` first")}const bt=n=>BigInt.asUintN(64,BigInt(n));function xs(n){return n>>>0}const In=new TextDecoder,wd=new TextEncoder;let wt=0;function tn(n,e,t){if(typeof n!="string")throw new TypeError("expected a string");if(n.length===0)return wt=0,1;let s=0,i=0,r=0;for(;n.length>0;){i=e(i,s,1,s+=n.length*2);const{read:o,written:a}=wd.encodeInto(n,new Uint8Array(t.buffer,i+r,s-r));r+=a,n=n.slice(o)}return wt=r,i}let Y,mt;function vd(){const n=ad();if(!(n instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var e=Ks++;return Qe.set(e,{rep:n,own:!0}),e}function xd(n){let e;switch(n){case 0:{e={tag:"ok",val:void 0};break}case 1:{e={tag:"err",val:void 0};break}default:throw new TypeError("invalid variant discriminant for expected")}od(e)}function kd(){const n=ld();if(!(n instanceof Qs))throw new Error('Resource error: Not a valid "InputStream" resource.');var e=Fa++;return ln.set(e,{rep:n,own:!0}),e}function Id(){const n=cd();if(!(n instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var e=Ks++;return Qe.set(e,{rep:n,own:!0}),e}let Be;function $d(n){var t=fd(),s=t.length,i=lt(0,0,4,s*12);for(let u=0;u<t.length;u++){const p=t[u],h=i+u*12;var[r,o]=p;if(!(r instanceof ct))throw new Error('Resource error: Not a valid "Descriptor" resource.');var a=ja++;He.set(a,{rep:r,own:!0}),k(v).setInt32(h+0,a,!0);var l=tn(o,lt,v),c=wt;k(v).setInt32(h+8,c,!0),k(v).setInt32(h+4,l,!0)}k(v).setInt32(n+4,s,!0),k(v).setInt32(n+0,i,!0)}let v,lt;function Sd(n,e,t){var s=n,i=He.get(s).rep;let r;try{r={tag:"ok",val:ct.prototype.readViaStream.call(i,BigInt.asUintN(64,e))}}catch(c){r={tag:"err",val:je(c)}}var o=r;switch(o.tag){case"ok":{const c=o.val;if(k(v).setInt8(t+0,0,!0),!(c instanceof Qs))throw new Error('Resource error: Not a valid "InputStream" resource.');var a=Fa++;ln.set(a,{rep:c,own:!0}),k(v).setInt32(t+4,a,!0);break}case"err":{const c=o.val;k(v).setInt8(t+0,1,!0);var l=c;let u;switch(l){case"access":{u=0;break}case"would-block":{u=1;break}case"already":{u=2;break}case"bad-descriptor":{u=3;break}case"busy":{u=4;break}case"deadlock":{u=5;break}case"quota":{u=6;break}case"exist":{u=7;break}case"file-too-large":{u=8;break}case"illegal-byte-sequence":{u=9;break}case"in-progress":{u=10;break}case"interrupted":{u=11;break}case"invalid":{u=12;break}case"io":{u=13;break}case"is-directory":{u=14;break}case"loop":{u=15;break}case"too-many-links":{u=16;break}case"message-size":{u=17;break}case"name-too-long":{u=18;break}case"no-device":{u=19;break}case"no-entry":{u=20;break}case"no-lock":{u=21;break}case"insufficient-memory":{u=22;break}case"insufficient-space":{u=23;break}case"not-directory":{u=24;break}case"not-empty":{u=25;break}case"not-recoverable":{u=26;break}case"unsupported":{u=27;break}case"no-tty":{u=28;break}case"no-such-device":{u=29;break}case"overflow":{u=30;break}case"not-permitted":{u=31;break}case"pipe":{u=32;break}case"read-only":{u=33;break}case"invalid-seek":{u=34;break}case"text-file-busy":{u=35;break}case"cross-device":{u=36;break}default:throw c instanceof Error&&console.error(c),new TypeError(`"${l}" is not one of the cases of error-code`)}k(v).setInt8(t+4,u,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Nd(n,e,t){var s=n,i=He.get(s).rep;let r;try{r={tag:"ok",val:ct.prototype.writeViaStream.call(i,BigInt.asUintN(64,e))}}catch(c){r={tag:"err",val:je(c)}}var o=r;switch(o.tag){case"ok":{const c=o.val;if(k(v).setInt8(t+0,0,!0),!(c instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var a=Ks++;Qe.set(a,{rep:c,own:!0}),k(v).setInt32(t+4,a,!0);break}case"err":{const c=o.val;k(v).setInt8(t+0,1,!0);var l=c;let u;switch(l){case"access":{u=0;break}case"would-block":{u=1;break}case"already":{u=2;break}case"bad-descriptor":{u=3;break}case"busy":{u=4;break}case"deadlock":{u=5;break}case"quota":{u=6;break}case"exist":{u=7;break}case"file-too-large":{u=8;break}case"illegal-byte-sequence":{u=9;break}case"in-progress":{u=10;break}case"interrupted":{u=11;break}case"invalid":{u=12;break}case"io":{u=13;break}case"is-directory":{u=14;break}case"loop":{u=15;break}case"too-many-links":{u=16;break}case"message-size":{u=17;break}case"name-too-long":{u=18;break}case"no-device":{u=19;break}case"no-entry":{u=20;break}case"no-lock":{u=21;break}case"insufficient-memory":{u=22;break}case"insufficient-space":{u=23;break}case"not-directory":{u=24;break}case"not-empty":{u=25;break}case"not-recoverable":{u=26;break}case"unsupported":{u=27;break}case"no-tty":{u=28;break}case"no-such-device":{u=29;break}case"overflow":{u=30;break}case"not-permitted":{u=31;break}case"pipe":{u=32;break}case"read-only":{u=33;break}case"invalid-seek":{u=34;break}case"text-file-busy":{u=35;break}case"cross-device":{u=36;break}default:throw c instanceof Error&&console.error(c),new TypeError(`"${l}" is not one of the cases of error-code`)}k(v).setInt8(t+4,u,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Pd(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.appendViaStream.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{const l=r.val;if(k(v).setInt8(e+0,0,!0),!(l instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var o=Ks++;Qe.set(o,{rep:l,own:!0}),k(v).setInt32(e+4,o,!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var a=l;let c;switch(a){case"access":{c=0;break}case"would-block":{c=1;break}case"already":{c=2;break}case"bad-descriptor":{c=3;break}case"busy":{c=4;break}case"deadlock":{c=5;break}case"quota":{c=6;break}case"exist":{c=7;break}case"file-too-large":{c=8;break}case"illegal-byte-sequence":{c=9;break}case"in-progress":{c=10;break}case"interrupted":{c=11;break}case"invalid":{c=12;break}case"io":{c=13;break}case"is-directory":{c=14;break}case"loop":{c=15;break}case"too-many-links":{c=16;break}case"message-size":{c=17;break}case"name-too-long":{c=18;break}case"no-device":{c=19;break}case"no-entry":{c=20;break}case"no-lock":{c=21;break}case"insufficient-memory":{c=22;break}case"insufficient-space":{c=23;break}case"not-directory":{c=24;break}case"not-empty":{c=25;break}case"not-recoverable":{c=26;break}case"unsupported":{c=27;break}case"no-tty":{c=28;break}case"no-such-device":{c=29;break}case"overflow":{c=30;break}case"not-permitted":{c=31;break}case"pipe":{c=32;break}case"read-only":{c=33;break}case"invalid-seek":{c=34;break}case"text-file-busy":{c=35;break}case"cross-device":{c=36;break}default:throw l instanceof Error&&console.error(l),new TypeError(`"${a}" is not one of the cases of error-code`)}k(v).setInt8(e+4,c,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Cd(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.getType.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{const l=r.val;k(v).setInt8(e+0,0,!0);var o=l;let c;switch(o){case"unknown":{c=0;break}case"block-device":{c=1;break}case"character-device":{c=2;break}case"directory":{c=3;break}case"fifo":{c=4;break}case"symbolic-link":{c=5;break}case"regular-file":{c=6;break}case"socket":{c=7;break}default:throw l instanceof Error&&console.error(l),new TypeError(`"${o}" is not one of the cases of descriptor-type`)}k(v).setInt8(e+1,c,!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var a=l;let c;switch(a){case"access":{c=0;break}case"would-block":{c=1;break}case"already":{c=2;break}case"bad-descriptor":{c=3;break}case"busy":{c=4;break}case"deadlock":{c=5;break}case"quota":{c=6;break}case"exist":{c=7;break}case"file-too-large":{c=8;break}case"illegal-byte-sequence":{c=9;break}case"in-progress":{c=10;break}case"interrupted":{c=11;break}case"invalid":{c=12;break}case"io":{c=13;break}case"is-directory":{c=14;break}case"loop":{c=15;break}case"too-many-links":{c=16;break}case"message-size":{c=17;break}case"name-too-long":{c=18;break}case"no-device":{c=19;break}case"no-entry":{c=20;break}case"no-lock":{c=21;break}case"insufficient-memory":{c=22;break}case"insufficient-space":{c=23;break}case"not-directory":{c=24;break}case"not-empty":{c=25;break}case"not-recoverable":{c=26;break}case"unsupported":{c=27;break}case"no-tty":{c=28;break}case"no-such-device":{c=29;break}case"overflow":{c=30;break}case"not-permitted":{c=31;break}case"pipe":{c=32;break}case"read-only":{c=33;break}case"invalid-seek":{c=34;break}case"text-file-busy":{c=35;break}case"cross-device":{c=36;break}default:throw l instanceof Error&&console.error(l),new TypeError(`"${a}" is not one of the cases of error-code`)}k(v).setInt8(e+1,c,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Md(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.stat.call(s)}}catch(P){i={tag:"err",val:je(P)}}var r=i;switch(r.tag){case"ok":{const P=r.val;k(v).setInt8(e+0,0,!0);var{type:o,linkCount:a,size:l,dataAccessTimestamp:c,dataModificationTimestamp:u,statusChangeTimestamp:p}=P,h=o;let S;switch(h){case"unknown":{S=0;break}case"block-device":{S=1;break}case"character-device":{S=2;break}case"directory":{S=3;break}case"fifo":{S=4;break}case"symbolic-link":{S=5;break}case"regular-file":{S=6;break}case"socket":{S=7;break}default:throw o instanceof Error&&console.error(o),new TypeError(`"${h}" is not one of the cases of descriptor-type`)}k(v).setInt8(e+8,S,!0),k(v).setBigInt64(e+16,bt(a),!0),k(v).setBigInt64(e+24,bt(l),!0);var f=c;if(f==null)k(v).setInt8(e+32,0,!0);else{const I=f;k(v).setInt8(e+32,1,!0);var{seconds:A,nanoseconds:m}=I;k(v).setBigInt64(e+40,bt(A),!0),k(v).setInt32(e+48,xs(m),!0)}var y=u;if(y==null)k(v).setInt8(e+56,0,!0);else{const I=y;k(v).setInt8(e+56,1,!0);var{seconds:b,nanoseconds:x}=I;k(v).setBigInt64(e+64,bt(b),!0),k(v).setInt32(e+72,xs(x),!0)}var g=p;if(g==null)k(v).setInt8(e+80,0,!0);else{const I=g;k(v).setInt8(e+80,1,!0);var{seconds:E,nanoseconds:w}=I;k(v).setBigInt64(e+88,bt(E),!0),k(v).setInt32(e+96,xs(w),!0)}break}case"err":{const P=r.val;k(v).setInt8(e+0,1,!0);var $=P;let S;switch($){case"access":{S=0;break}case"would-block":{S=1;break}case"already":{S=2;break}case"bad-descriptor":{S=3;break}case"busy":{S=4;break}case"deadlock":{S=5;break}case"quota":{S=6;break}case"exist":{S=7;break}case"file-too-large":{S=8;break}case"illegal-byte-sequence":{S=9;break}case"in-progress":{S=10;break}case"interrupted":{S=11;break}case"invalid":{S=12;break}case"io":{S=13;break}case"is-directory":{S=14;break}case"loop":{S=15;break}case"too-many-links":{S=16;break}case"message-size":{S=17;break}case"name-too-long":{S=18;break}case"no-device":{S=19;break}case"no-entry":{S=20;break}case"no-lock":{S=21;break}case"insufficient-memory":{S=22;break}case"insufficient-space":{S=23;break}case"not-directory":{S=24;break}case"not-empty":{S=25;break}case"not-recoverable":{S=26;break}case"unsupported":{S=27;break}case"no-tty":{S=28;break}case"no-such-device":{S=29;break}case"overflow":{S=30;break}case"not-permitted":{S=31;break}case"pipe":{S=32;break}case"read-only":{S=33;break}case"invalid-seek":{S=34;break}case"text-file-busy":{S=35;break}case"cross-device":{S=36;break}default:throw P instanceof Error&&console.error(P),new TypeError(`"${$}" is not one of the cases of error-code`)}k(v).setInt8(e+8,S,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Rd(n,e,t,s,i,r,o){var a=n,l=He.get(a).rep;if(e&4294967294)throw new TypeError("flags have extraneous bits set");var c={symlinkFollow:!!(e&1)},u=t,p=s,h=In.decode(new Uint8Array(v.buffer,u,p));if(i&4294967280)throw new TypeError("flags have extraneous bits set");var f={create:!!(i&1),directory:!!(i&2),exclusive:!!(i&4),truncate:!!(i&8)};if(r&4294967232)throw new TypeError("flags have extraneous bits set");var A={read:!!(r&1),write:!!(r&2),fileIntegritySync:!!(r&4),dataIntegritySync:!!(r&8),requestedWriteSync:!!(r&16),mutateDirectory:!!(r&32)};let m;try{m={tag:"ok",val:ct.prototype.openAt.call(l,c,h,f,A)}}catch(g){m={tag:"err",val:je(g)}}var y=m;switch(y.tag){case"ok":{const g=y.val;if(k(v).setInt8(o+0,0,!0),!(g instanceof ct))throw new Error('Resource error: Not a valid "Descriptor" resource.');var b=ja++;He.set(b,{rep:g,own:!0}),k(v).setInt32(o+4,b,!0);break}case"err":{const g=y.val;k(v).setInt8(o+0,1,!0);var x=g;let E;switch(x){case"access":{E=0;break}case"would-block":{E=1;break}case"already":{E=2;break}case"bad-descriptor":{E=3;break}case"busy":{E=4;break}case"deadlock":{E=5;break}case"quota":{E=6;break}case"exist":{E=7;break}case"file-too-large":{E=8;break}case"illegal-byte-sequence":{E=9;break}case"in-progress":{E=10;break}case"interrupted":{E=11;break}case"invalid":{E=12;break}case"io":{E=13;break}case"is-directory":{E=14;break}case"loop":{E=15;break}case"too-many-links":{E=16;break}case"message-size":{E=17;break}case"name-too-long":{E=18;break}case"no-device":{E=19;break}case"no-entry":{E=20;break}case"no-lock":{E=21;break}case"insufficient-memory":{E=22;break}case"insufficient-space":{E=23;break}case"not-directory":{E=24;break}case"not-empty":{E=25;break}case"not-recoverable":{E=26;break}case"unsupported":{E=27;break}case"no-tty":{E=28;break}case"no-such-device":{E=29;break}case"overflow":{E=30;break}case"not-permitted":{E=31;break}case"pipe":{E=32;break}case"read-only":{E=33;break}case"invalid-seek":{E=34;break}case"text-file-busy":{E=35;break}case"cross-device":{E=36;break}default:throw g instanceof Error&&console.error(g),new TypeError(`"${x}" is not one of the cases of error-code`)}k(v).setInt8(o+4,E,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Ld(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.metadataHash.call(s)}}catch(c){i={tag:"err",val:je(c)}}var r=i;switch(r.tag){case"ok":{const c=r.val;k(v).setInt8(e+0,0,!0);var{lower:o,upper:a}=c;k(v).setBigInt64(e+8,bt(o),!0),k(v).setBigInt64(e+16,bt(a),!0);break}case"err":{const c=r.val;k(v).setInt8(e+0,1,!0);var l=c;let u;switch(l){case"access":{u=0;break}case"would-block":{u=1;break}case"already":{u=2;break}case"bad-descriptor":{u=3;break}case"busy":{u=4;break}case"deadlock":{u=5;break}case"quota":{u=6;break}case"exist":{u=7;break}case"file-too-large":{u=8;break}case"illegal-byte-sequence":{u=9;break}case"in-progress":{u=10;break}case"interrupted":{u=11;break}case"invalid":{u=12;break}case"io":{u=13;break}case"is-directory":{u=14;break}case"loop":{u=15;break}case"too-many-links":{u=16;break}case"message-size":{u=17;break}case"name-too-long":{u=18;break}case"no-device":{u=19;break}case"no-entry":{u=20;break}case"no-lock":{u=21;break}case"insufficient-memory":{u=22;break}case"insufficient-space":{u=23;break}case"not-directory":{u=24;break}case"not-empty":{u=25;break}case"not-recoverable":{u=26;break}case"unsupported":{u=27;break}case"no-tty":{u=28;break}case"no-such-device":{u=29;break}case"overflow":{u=30;break}case"not-permitted":{u=31;break}case"pipe":{u=32;break}case"read-only":{u=33;break}case"invalid-seek":{u=34;break}case"text-file-busy":{u=35;break}case"cross-device":{u=36;break}default:throw c instanceof Error&&console.error(c),new TypeError(`"${l}" is not one of the cases of error-code`)}k(v).setInt8(e+8,u,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Od(n,e){var t=n,s=dt.get(t).rep,r=md(s);if(r==null)k(v).setInt8(e+0,0,!0);else{const a=r;k(v).setInt8(e+0,1,!0);var o=a;let l;switch(o){case"access":{l=0;break}case"would-block":{l=1;break}case"already":{l=2;break}case"bad-descriptor":{l=3;break}case"busy":{l=4;break}case"deadlock":{l=5;break}case"quota":{l=6;break}case"exist":{l=7;break}case"file-too-large":{l=8;break}case"illegal-byte-sequence":{l=9;break}case"in-progress":{l=10;break}case"interrupted":{l=11;break}case"invalid":{l=12;break}case"io":{l=13;break}case"is-directory":{l=14;break}case"loop":{l=15;break}case"too-many-links":{l=16;break}case"message-size":{l=17;break}case"name-too-long":{l=18;break}case"no-device":{l=19;break}case"no-entry":{l=20;break}case"no-lock":{l=21;break}case"insufficient-memory":{l=22;break}case"insufficient-space":{l=23;break}case"not-directory":{l=24;break}case"not-empty":{l=25;break}case"not-recoverable":{l=26;break}case"unsupported":{l=27;break}case"no-tty":{l=28;break}case"no-such-device":{l=29;break}case"overflow":{l=30;break}case"not-permitted":{l=31;break}case"pipe":{l=32;break}case"read-only":{l=33;break}case"invalid-seek":{l=34;break}case"text-file-busy":{l=35;break}case"cross-device":{l=36;break}default:throw a instanceof Error&&console.error(a),new TypeError(`"${o}" is not one of the cases of error-code`)}k(v).setInt8(e+1,l,!0)}}function Dd(n,e,t){var s=n,i=ln.get(s).rep;let r;try{r={tag:"ok",val:Qs.prototype.read.call(i,BigInt.asUintN(64,e))}}catch(f){r={tag:"err",val:je(f)}}var o=r;switch(o.tag){case"ok":{const f=o.val;k(v).setInt8(t+0,0,!0);var a=f,l=a.byteLength,c=lt(0,0,1,l*1),u=new Uint8Array(a.buffer||a,a.byteOffset,l*1);new Uint8Array(v.buffer,c,l*1).set(u),k(v).setInt32(t+8,l,!0),k(v).setInt32(t+4,c,!0);break}case"err":{const f=o.val;k(v).setInt8(t+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(t+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(t+8,h,!0);break}case"closed":{k(v).setInt8(t+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Td(n,e,t){var s=n,i=ln.get(s).rep;let r;try{r={tag:"ok",val:Qs.prototype.blockingRead.call(i,BigInt.asUintN(64,e))}}catch(f){r={tag:"err",val:je(f)}}var o=r;switch(o.tag){case"ok":{const f=o.val;k(v).setInt8(t+0,0,!0);var a=f,l=a.byteLength,c=lt(0,0,1,l*1),u=new Uint8Array(a.buffer||a,a.byteOffset,l*1);new Uint8Array(v.buffer,c,l*1).set(u),k(v).setInt32(t+8,l,!0),k(v).setInt32(t+4,c,!0);break}case"err":{const f=o.val;k(v).setInt8(t+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(t+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(t+8,h,!0);break}case"closed":{k(v).setInt8(t+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Bd(n,e){var t=n,s=Qe.get(t).rep;let i;try{i={tag:"ok",val:Nt.prototype.checkWrite.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{const l=r.val;k(v).setInt8(e+0,0,!0),k(v).setBigInt64(e+8,bt(l),!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var o=l;switch(o.tag){case"last-operation-failed":{const c=o.val;if(k(v).setInt8(e+8,0,!0),!(c instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var a=hn++;dt.set(a,{rep:c,own:!0}),k(v).setInt32(e+12,a,!0);break}case"closed":{k(v).setInt8(e+8,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(o.tag)}\` (received \`${o}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function _d(n,e,t,s){var i=n,r=Qe.get(i).rep,o=e,a=t,l=new Uint8Array(v.buffer.slice(o,o+a*1));let c;try{c={tag:"ok",val:Nt.prototype.write.call(r,l)}}catch(f){c={tag:"err",val:je(f)}}var u=c;switch(u.tag){case"ok":{u.val,k(v).setInt8(s+0,0,!0);break}case"err":{const f=u.val;k(v).setInt8(s+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(s+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(s+8,h,!0);break}case"closed":{k(v).setInt8(s+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function zd(n,e,t,s){var i=n,r=Qe.get(i).rep,o=e,a=t,l=new Uint8Array(v.buffer.slice(o,o+a*1));let c;try{c={tag:"ok",val:Nt.prototype.blockingWriteAndFlush.call(r,l)}}catch(f){c={tag:"err",val:je(f)}}var u=c;switch(u.tag){case"ok":{u.val,k(v).setInt8(s+0,0,!0);break}case"err":{const f=u.val;k(v).setInt8(s+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(s+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(s+8,h,!0);break}case"closed":{k(v).setInt8(s+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Vd(n,e){var t=n,s=Qe.get(t).rep;let i;try{i={tag:"ok",val:Nt.prototype.blockingFlush.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{r.val,k(v).setInt8(e+0,0,!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var o=l;switch(o.tag){case"last-operation-failed":{const c=o.val;if(k(v).setInt8(e+4,0,!0),!(c instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var a=hn++;dt.set(a,{rep:c,own:!0}),k(v).setInt32(e+8,a,!0);break}case"closed":{k(v).setInt8(e+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(o.tag)}\` (received \`${o}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Fd(n,e){var s=gd(BigInt.asUintN(64,n)),i=s.byteLength,r=lt(0,0,1,i*1),o=new Uint8Array(s.buffer||s,s.byteOffset,i*1);new Uint8Array(v.buffer,r,i*1).set(o),k(v).setInt32(e+4,i,!0),k(v).setInt32(e+0,r,!0)}function jd(n){var t=rd(),s=t.length,i=lt(0,0,4,s*16);for(let p=0;p<t.length;p++){const h=t[p],f=i+p*16;var[r,o]=h,a=tn(r,lt,v),l=wt;k(v).setInt32(f+4,l,!0),k(v).setInt32(f+0,a,!0);var c=tn(o,lt,v),u=wt;k(v).setInt32(f+12,u,!0),k(v).setInt32(f+8,c,!0)}k(v).setInt32(n+4,s,!0),k(v).setInt32(n+0,i,!0)}function Ud(n){var t=hd();if(t==null)k(v).setInt8(n+0,0,!0);else{const i=t;if(k(v).setInt8(n+0,1,!0),!(i instanceof dd))throw new Error('Resource error: Not a valid "TerminalInput" resource.');var s=tu++;Ji.set(s,{rep:i,own:!0}),k(v).setInt32(n+4,s,!0)}}function Gd(n){var t=pd();if(t==null)k(v).setInt8(n+0,0,!0);else{const i=t;if(k(v).setInt8(n+0,1,!0),!(i instanceof za))throw new Error('Resource error: Not a valid "TerminalOutput" resource.');var s=Ua++;Rs.set(s,{rep:i,own:!0}),k(v).setInt32(n+4,s,!0)}}function Wd(n){var t=ud();if(t==null)k(v).setInt8(n+0,0,!0);else{const i=t;if(k(v).setInt8(n+0,1,!0),!(i instanceof za))throw new Error('Resource error: Not a valid "TerminalOutput" resource.');var s=Ua++;Rs.set(s,{rep:i,own:!0}),k(v).setInt32(n+4,s,!0)}}let Ct,Va;function Hd(n){const e=lo.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);lo.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function qd(n){const e=dt.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);dt.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Xd(n){const e=ln.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);ln.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Zd(n){const e=Qe.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);Qe.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Yd(n){const e=He.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);He.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Qd(n){const e=co.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);co.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Kd(n){const e=Rs.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);Rs.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Jd(n){const e=Ji.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);Ji.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function eu(n,e){Ga||Ad();var t=Ct(0,0,4,52),s=n,i=s.byteLength,r=Ct(0,0,1,i*1),o=new Uint8Array(s.buffer||s,s.byteOffset,i*1);new Uint8Array(v.buffer,r,i*1).set(o),k(v).setInt32(t+4,i,!0),k(v).setInt32(t+0,r,!0);var{name:a,noTypescript:l,instantiation:c,map:u,compat:p,noNodejsCompat:h,base64Cutoff:f,tlaCompat:A,validLiftingOptimization:m,tracing:y,noNamespacedExports:b}=e,x=tn(a,Ct,v),g=wt;k(v).setInt32(t+12,g,!0),k(v).setInt32(t+8,x,!0);var E=l;if(E==null)k(v).setInt8(t+16,0,!0);else{const Q=E;k(v).setInt8(t+16,1,!0),k(v).setInt8(t+17,Q?1:0,!0)}var w=c;if(w==null)k(v).setInt8(t+18,0,!0);else{const Q=w;k(v).setInt8(t+18,1,!0);var $=Q;switch($.tag){case"async":{k(v).setInt8(t+19,0,!0);break}case"sync":{k(v).setInt8(t+19,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify($.tag)}\` (received \`${$}\`) specified for \`InstantiationMode\``)}}var P=u;if(P==null)k(v).setInt8(t+20,0,!0);else{const Q=P;k(v).setInt8(t+20,1,!0);var S=Q,I=S.length,C=Ct(0,0,4,I*16);for(let Ce=0;Ce<S.length;Ce++){const An=S[Ce],os=C+Ce*16;var[V,R]=An,_=tn(V,Ct,v),D=wt;k(v).setInt32(os+4,D,!0),k(v).setInt32(os+0,_,!0);var L=tn(R,Ct,v),B=wt;k(v).setInt32(os+12,B,!0),k(v).setInt32(os+8,L,!0)}k(v).setInt32(t+28,I,!0),k(v).setInt32(t+24,C,!0)}var F=p;if(F==null)k(v).setInt8(t+32,0,!0);else{const Q=F;k(v).setInt8(t+32,1,!0),k(v).setInt8(t+33,Q?1:0,!0)}var H=h;if(H==null)k(v).setInt8(t+34,0,!0);else{const Q=H;k(v).setInt8(t+34,1,!0),k(v).setInt8(t+35,Q?1:0,!0)}var oe=f;if(oe==null)k(v).setInt8(t+36,0,!0);else{const Q=oe;k(v).setInt8(t+36,1,!0),k(v).setInt32(t+40,xs(Q),!0)}var K=A;if(K==null)k(v).setInt8(t+44,0,!0);else{const Q=K;k(v).setInt8(t+44,1,!0),k(v).setInt8(t+45,Q?1:0,!0)}var G=m;if(G==null)k(v).setInt8(t+46,0,!0);else{const Q=G;k(v).setInt8(t+46,1,!0),k(v).setInt8(t+47,Q?1:0,!0)}var ae=y;if(ae==null)k(v).setInt8(t+48,0,!0);else{const Q=ae;k(v).setInt8(t+48,1,!0),k(v).setInt8(t+49,Q?1:0,!0)}var Ne=b;if(Ne==null)k(v).setInt8(t+50,0,!0);else{const Q=Ne;k(v).setInt8(t+50,1,!0),k(v).setInt8(t+51,Q?1:0,!0)}const we=mt.generate(t);let he;switch(k(v).getUint8(we+0,!0)){case 0:{var Pe=k(v).getInt32(we+8,!0),tt=k(v).getInt32(we+4,!0),ve=[];for(let Q=0;Q<Pe;Q++){const Ce=tt+Q*16;var ge=k(v).getInt32(Ce+0,!0),pt=k(v).getInt32(Ce+4,!0),yn=In.decode(new Uint8Array(v.buffer,ge,pt)),is=k(v).getInt32(Ce+8,!0),pi=k(v).getInt32(Ce+12,!0),fi=new Uint8Array(v.buffer.slice(is,is+pi*1));ve.push([yn,fi])}var mi=k(v).getInt32(we+16,!0),gi=k(v).getInt32(we+12,!0),rs=[];for(let Q=0;Q<mi;Q++){const Ce=gi+Q*8;var Gt=k(v).getInt32(Ce+0,!0),Re=k(v).getInt32(Ce+4,!0),bn=In.decode(new Uint8Array(v.buffer,Gt,Re));rs.push(bn)}var yi=k(v).getInt32(we+24,!0),bi=k(v).getInt32(we+20,!0),En=[];for(let Q=0;Q<yi;Q++){const Ce=bi+Q*12;var Ei=k(v).getInt32(Ce+0,!0),Ai=k(v).getInt32(Ce+4,!0),wi=In.decode(new Uint8Array(v.buffer,Ei,Ai));let An;switch(k(v).getUint8(Ce+8,!0)){case 0:{An="function";break}case 1:{An="instance";break}default:throw new TypeError("invalid discriminant specified for ExportType")}En.push([wi,An])}he={tag:"ok",val:{files:ve,imports:rs,exports:En}};break}case 1:{var vc=k(v).getInt32(we+4,!0),xc=k(v).getInt32(we+8,!0),kc=In.decode(new Uint8Array(v.buffer,vc,xc));he={tag:"err",val:kc};break}default:throw new TypeError("invalid variant discriminant for expected")}if(Va(we),he.tag==="err")throw new yd(he.val);return he.val}const dt=new Map;let hn=0;const ln=new Map;let Fa=0;const Qe=new Map;let Ks=0;const He=new Map;let ja=0;const lo=new Map,co=new Map,Ji=new Map;let tu=0;const Rs=new Map;let Ua=0,Ga=!1;const nu=(async()=>{const n=ao(new URL("/loader/assets/js-component-bindgen-component.core-8p6hfTFd.wasm",import.meta.url)),e=ao(new URL("/loader/assets/js-component-bindgen-component.core2-0P6l7xr-.wasm",import.meta.url)),t=oo("AGFzbQEAAAABUQxgAX8AYAN/fn8AYAJ/fwBgB39/f39/f38AYAR/f39/AGACfn8AYAl/f39/f35+f38Bf2ACf38Bf2AEf39/fwF/YAF/AX9gA39/fwF/YAF/AAMgHwABAQICAgMCAgEBAgQEAgUAAAAABgcHCAgHBwkHCgsEBQFwAR8fB50BIAEwAAABMQABATIAAgEzAAMBNAAEATUABQE2AAYBNwAHATgACAE5AAkCMTAACgIxMQALAjEyAAwCMTMADQIxNAAOAjE1AA8CMTYAEAIxNwARAjE4ABICMTkAEwIyMAAUAjIxABUCMjIAFgIyMwAXAjI0ABgCMjUAGQIyNgAaAjI3ABsCMjgAHAIyOQAdAjMwAB4IJGltcG9ydHMBAAqZAx8JACAAQQARAAALDQAgACABIAJBAREBAAsNACAAIAEgAkECEQEACwsAIAAgAUEDEQIACwsAIAAgAUEEEQIACwsAIAAgAUEFEQIACxUAIAAgASACIAMgBCAFIAZBBhEDAAsLACAAIAFBBxECAAsLACAAIAFBCBECAAsNACAAIAEgAkEJEQEACw0AIAAgASACQQoRAQALCwAgACABQQsRAgALDwAgACABIAIgA0EMEQQACw8AIAAgASACIANBDREEAAsLACAAIAFBDhECAAsLACAAIAFBDxEFAAsJACAAQRARAAALCQAgAEEREQAACwkAIABBEhEAAAsJACAAQRMRAAALGQAgACABIAIgAyAEIAUgBiAHIAhBFBEGAAsLACAAIAFBFREHAAsLACAAIAFBFhEHAAsPACAAIAEgAiADQRcRCAALDwAgACABIAIgA0EYEQgACwsAIAAgAUEZEQcACwsAIAAgAUEaEQcACwkAIABBGxEJAAsLACAAIAFBHBEHAAsNACAAIAEgAkEdEQoACwkAIABBHhELAAsALglwcm9kdWNlcnMBDHByb2Nlc3NlZC1ieQENd2l0LWNvbXBvbmVudAYwLjE5LjAAkhAEbmFtZQATEndpdC1jb21wb25lbnQ6c2hpbQH1Dx8ARWluZGlyZWN0LXdhc2k6ZmlsZXN5c3RlbS9wcmVvcGVuc0AwLjIuMC1yYy0yMDIzLTExLTEwLWdldC1kaXJlY3RvcmllcwFVaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLnJlYWQtdmlhLXN0cmVhbQJWaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLndyaXRlLXZpYS1zdHJlYW0DV2luZGlyZWN0LXdhc2k6ZmlsZXN5c3RlbS90eXBlc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdZGVzY3JpcHRvci5hcHBlbmQtdmlhLXN0cmVhbQROaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLmdldC10eXBlBUppbmRpcmVjdC13YXNpOmZpbGVzeXN0ZW0vdHlwZXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1bbWV0aG9kXWRlc2NyaXB0b3Iuc3RhdAZNaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLm9wZW4tYXQHU2luZGlyZWN0LXdhc2k6ZmlsZXN5c3RlbS90eXBlc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdZGVzY3JpcHRvci5tZXRhZGF0YS1oYXNoCEhpbmRpcmVjdC13YXNpOmZpbGVzeXN0ZW0vdHlwZXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1maWxlc3lzdGVtLWVycm9yLWNvZGUJRmluZGlyZWN0LXdhc2k6aW8vc3RyZWFtc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdaW5wdXQtc3RyZWFtLnJlYWQKT2luZGlyZWN0LXdhc2k6aW8vc3RyZWFtc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdaW5wdXQtc3RyZWFtLmJsb2NraW5nLXJlYWQLTmluZGlyZWN0LXdhc2k6aW8vc3RyZWFtc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2Rdb3V0cHV0LXN0cmVhbS5jaGVjay13cml0ZQxIaW5kaXJlY3Qtd2FzaTppby9zdHJlYW1zQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1vdXRwdXQtc3RyZWFtLndyaXRlDVtpbmRpcmVjdC13YXNpOmlvL3N0cmVhbXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1bbWV0aG9kXW91dHB1dC1zdHJlYW0uYmxvY2tpbmctd3JpdGUtYW5kLWZsdXNoDlFpbmRpcmVjdC13YXNpOmlvL3N0cmVhbXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1bbWV0aG9kXW91dHB1dC1zdHJlYW0uYmxvY2tpbmctZmx1c2gPQGluZGlyZWN0LXdhc2k6cmFuZG9tL3JhbmRvbUAwLjIuMC1yYy0yMDIzLTExLTEwLWdldC1yYW5kb20tYnl0ZXMQQWluZGlyZWN0LXdhc2k6Y2xpL2Vudmlyb25tZW50QDAuMi4wLXJjLTIwMjMtMTEtMTAtZ2V0LWVudmlyb25tZW50EUdpbmRpcmVjdC13YXNpOmNsaS90ZXJtaW5hbC1zdGRpbkAwLjIuMC1yYy0yMDIzLTExLTEwLWdldC10ZXJtaW5hbC1zdGRpbhJJaW5kaXJlY3Qtd2FzaTpjbGkvdGVybWluYWwtc3Rkb3V0QDAuMi4wLXJjLTIwMjMtMTEtMTAtZ2V0LXRlcm1pbmFsLXN0ZG91dBNJaW5kaXJlY3Qtd2FzaTpjbGkvdGVybWluYWwtc3RkZXJyQDAuMi4wLXJjLTIwMjMtMTEtMTAtZ2V0LXRlcm1pbmFsLXN0ZGVychQmYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1wYXRoX29wZW4VLGFkYXB0LXdhc2lfc25hcHNob3RfcHJldmlldzEtZmRfZmlsZXN0YXRfZ2V0FidhZGFwdC13YXNpX3NuYXBzaG90X3ByZXZpZXcxLXJhbmRvbV9nZXQXJGFkYXB0LXdhc2lfc25hcHNob3RfcHJldmlldzEtZmRfcmVhZBglYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF93cml0ZRkoYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1lbnZpcm9uX2dldBouYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1lbnZpcm9uX3NpemVzX2dldBslYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF9jbG9zZRwrYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF9wcmVzdGF0X2dldB0wYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF9wcmVzdGF0X2Rpcl9uYW1lHiZhZGFwdC13YXNpX3NuYXBzaG90X3ByZXZpZXcxLXByb2NfZXhpdA"),s=oo("AGFzbQEAAAABUQxgAX8AYAN/fn8AYAJ/fwBgB39/f39/f38AYAR/f39/AGACfn8AYAl/f39/f35+f38Bf2ACf38Bf2AEf39/fwF/YAF/AX9gA39/fwF/YAF/AALAASAAATAAAAABMQABAAEyAAEAATMAAgABNAACAAE1AAIAATYAAwABNwACAAE4AAIAATkAAQACMTAAAQACMTEAAgACMTIABAACMTMABAACMTQAAgACMTUABQACMTYAAAACMTcAAAACMTgAAAACMTkAAAACMjAABgACMjEABwACMjIABwACMjMACAACMjQACAACMjUABwACMjYABwACMjcACQACMjgABwACMjkACgACMzAACwAIJGltcG9ydHMBcAEfHwklAQBBAAsfAAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHgAuCXByb2R1Y2VycwEMcHJvY2Vzc2VkLWJ5AQ13aXQtY29tcG9uZW50BjAuMTkuMAAcBG5hbWUAFRR3aXQtY29tcG9uZW50OmZpeHVwcw");({exports:Y}=await ds(await t)),{exports:mt}=await ds(await n,{wasi_snapshot_preview1:{environ_get:Y[25],environ_sizes_get:Y[26],fd_close:Y[27],fd_filestat_get:Y[21],fd_prestat_dir_name:Y[29],fd_prestat_get:Y[28],fd_read:Y[23],fd_write:Y[24],path_open:Y[20],proc_exit:Y[30],random_get:Y[22]}}),{exports:Be}=await ds(await e,{__main_module__:{cabi_realloc:mt.cabi_realloc},env:{memory:mt.memory},"wasi:cli/environment@0.2.0-rc-2023-11-10":{"get-environment":Y[16]},"wasi:cli/exit@0.2.0-rc-2023-11-10":{exit:xd},"wasi:cli/stderr@0.2.0-rc-2023-11-10":{"get-stderr":vd},"wasi:cli/stdin@0.2.0-rc-2023-11-10":{"get-stdin":kd},"wasi:cli/stdout@0.2.0-rc-2023-11-10":{"get-stdout":Id},"wasi:cli/terminal-input@0.2.0-rc-2023-11-10":{"[resource-drop]terminal-input":Jd},"wasi:cli/terminal-output@0.2.0-rc-2023-11-10":{"[resource-drop]terminal-output":Kd},"wasi:cli/terminal-stderr@0.2.0-rc-2023-11-10":{"get-terminal-stderr":Y[19]},"wasi:cli/terminal-stdin@0.2.0-rc-2023-11-10":{"get-terminal-stdin":Y[17]},"wasi:cli/terminal-stdout@0.2.0-rc-2023-11-10":{"get-terminal-stdout":Y[18]},"wasi:filesystem/preopens@0.2.0-rc-2023-11-10":{"get-directories":Y[0]},"wasi:filesystem/types@0.2.0-rc-2023-11-10":{"[method]descriptor.append-via-stream":Y[3],"[method]descriptor.get-type":Y[4],"[method]descriptor.metadata-hash":Y[7],"[method]descriptor.open-at":Y[6],"[method]descriptor.read-via-stream":Y[1],"[method]descriptor.stat":Y[5],"[method]descriptor.write-via-stream":Y[2],"[resource-drop]descriptor":Yd,"[resource-drop]directory-entry-stream":Hd,"filesystem-error-code":Y[8]},"wasi:io/error@0.2.0-rc-2023-11-10":{"[resource-drop]error":qd},"wasi:io/streams@0.2.0-rc-2023-11-10":{"[method]input-stream.blocking-read":Y[10],"[method]input-stream.read":Y[9],"[method]output-stream.blocking-flush":Y[14],"[method]output-stream.blocking-write-and-flush":Y[13],"[method]output-stream.check-write":Y[11],"[method]output-stream.write":Y[12],"[resource-drop]input-stream":Xd,"[resource-drop]output-stream":Zd},"wasi:random/random@0.2.0-rc-2023-11-10":{"get-random-bytes":Y[15]},"wasi:sockets/tcp@0.2.0-rc-2023-11-10":{"[resource-drop]tcp-socket":Qd}}),v=mt.memory,lt=Be.cabi_import_realloc,await ds(await s,{"":{$imports:Y.$imports,0:$d,1:Sd,10:Td,11:Bd,12:_d,13:zd,14:Vd,15:Fd,16:jd,17:Ud,18:Gd,19:Wd,2:Nd,20:Be.path_open,21:Be.fd_filestat_get,22:Be.random_get,23:Be.fd_read,24:Be.fd_write,25:Be.environ_get,26:Be.environ_sizes_get,27:Be.fd_close,28:Be.fd_prestat_get,29:Be.fd_prestat_dir_name,3:Pd,30:Be.proc_exit,4:Cd,5:Md,6:Rd,7:Ld,8:Od,9:Dd}}),Ct=mt.cabi_realloc,Va=mt.cabi_post_generate,mt["cabi_post_generate-types"],Ga=!0})();async function su(){return await nu,eu.apply(this,arguments)}/*
+var no=(n,e,t)=>{if(!e.has(n))throw TypeError("Cannot "+t)};var xe=(n,e,t)=>(no(n,e,"read from private field"),t?t.call(n):e.get(n)),as=(n,e,t)=>{if(e.has(n))throw TypeError("Cannot add the same private member more than once");e instanceof WeakSet?e.add(n):e.set(n,t)},vi=(n,e,t,s)=>(no(n,e,"write to private field"),s?s.call(n,t):e.set(n,t),t);import{_ as Ra}from"./index-IMu5n1Ph.js";function Ic(n){return new Promise(e=>{setTimeout(e,n)})}let so="https://unpkg.com";const ls=new Map;async function Qi(n){if(ls.has(n))return ls.get(n);await Ic(200);const e=fetch(n).then(async t=>{if(t.ok)return{url:t.url,body:await t.text()};throw new Error(await t.text())}).catch(t=>{throw ls.delete(n),t});return ls.set(n,e),e}async function xi(n){return(await Qi(n)).url}const $c=new TextDecoder,Sc=n=>({name:"loader",async resolveId(e,t){if(n.find(s=>s[0]==e)||(e.endsWith("/")&&(e=e.slice(0,-1)),e.startsWith("http:")||e.startsWith("https:")))return e;if(e.startsWith(".")){const s=new URL(e,t).href;return await xi(s)}else{try{const s=await xi(`${so}/${e}/package.json`),i=(await Qi(s)).body,r=JSON.parse(i);if(r.svelte||r.module||r.main){const o=s.replace(/\/package\.json$/,"");return new URL(r.svelte||r.module||r.main,`${o}/`).href}}catch{}return await xi(`${so}/${e}`)}},async load(e){let t=n.find(i=>i[0]==e);return t?typeof t[1]=="string"?t[1]:$c.decode(t[1]):(await Qi(e)).body},async transform(e,t){if(!t.endsWith(".js"))return;let s=e.match(/new URL\((.*)/g);if(!(!s||!s.length))return s.forEach(i=>{let r=i.match(/'.\/(.*)'/)[1];if(!r.endsWith(".wasm"))return;let o=n.find(l=>l[0]==r);if(!o)return;let a=URL.createObjectURL(new Blob([o[1]],{type:"application/wasm"}));!a||a=="undefined"||(e=e.replace(`new URL('./${r}', import.meta.url)`,`'${a}'`))}),{code:e,map:{mappings:""}}}});let La=0;const Oa=Symbol.dispose||Symbol.for("dispose"),Nc=class{constructor(e){this.msg=e}toDebugString(){return this.msg}};let Pc=class{constructor(e){e||console.trace("no handler"),this.id=++La,this.handler=e}read(e){return this.handler.read?this.handler.read(e):this.handler.blockingRead.call(this,e)}blockingRead(e){return this.handler.blockingRead.call(this,e)}skip(e){if(this.handler.skip)return this.handler.skip.call(this,e);if(this.handler.read){const t=this.handler.read.call(this,e);return BigInt(t.byteLength)}return this.blockingSkip.call(this,e)}blockingSkip(e){if(this.handler.blockingSkip)return this.handler.blockingSkip.call(this,e);const t=this.handler.blockingRead.call(this,e);return BigInt(t.byteLength)}subscribe(){console.log(`[streams] Subscribe to input stream ${this.id}`)}[Oa](){this.handler.drop&&this.handler.drop.call(this)}},Cc=class{constructor(e){e||console.trace("no handler"),this.id=++La,this.open=!0,this.handler=e}checkWrite(e){return this.open?this.handler.checkWrite?this.handler.checkWrite.call(this,e):1000000n:0n}write(e){this.handler.write.call(this,e)}blockingWriteAndFlush(e){this.handler.write.call(this,e)}flush(){this.handler.flush&&this.handler.flush.call(this)}blockingFlush(){this.open=!0}writeZeroes(e){this.write.call(this,new Uint8Array(Number(e)))}blockingWriteZeroes(e){this.blockingWrite.call(this,new Uint8Array(Number(e)))}blockingWriteZeroesAndFlush(e){this.blockingWriteAndFlush.call(this,new Uint8Array(Number(e)))}splice(e,t){const s=Math.min(t,this.checkWrite.call(this)),i=e.read(s);return this.write.call(this,i),i.byteLength}blockingSplice(e,t){console.log(`[streams] Blocking splice ${this.id}`)}forward(e){console.log(`[streams] Forward ${this.id}`)}subscribe(){console.log(`[streams] Subscribe to output stream ${this.id}`)}[Oa](){}};const Mc={Error:Nc},Ir={InputStream:Pc,OutputStream:Cc},{InputStream:Rc,OutputStream:Lc}=Ir;let Oc=null,Dc={dir:{}};const Wt={seconds:BigInt(0),nanoseconds:0};function ki(n,e,t){e==="."&&io&&Tc(io[0])===n&&(e=Oc,e.startsWith("/")&&e!=="/"&&(e=e.slice(1)));let s=n,i;do{if(!s||!s.dir)throw"not-directory";i=e.indexOf("/");const r=i===-1?e:e.slice(0,i);if(r==="."||r==="")return s;if(r==="..")throw"no-entry";!s.dir[r]&&t.create?s=s.dir[r]=t.directory?{dir:{}}:{source:new Uint8Array([])}:s=s.dir[r]}while(i!==-1);if(!s)throw"no-entry";return s}function cs(n){return typeof n.source=="string"&&(n.source=new TextEncoder().encode(n.source)),n.source}let Da=class{constructor(e){this.idx=0,this.entries=e}readDirectoryEntry(){if(this.idx===this.entries.length)return null;const[e,t]=this.entries[this.idx];return this.idx+=1,{name:e,type:t.dir?"directory":"regular-file"}}};var qn,ke,Xn,Zn;let Zs=(Zn=class{constructor(e,t){as(this,qn,void 0);as(this,ke,void 0);as(this,Xn,0);t?vi(this,qn,e):vi(this,ke,e)}_getEntry(e){return xe(e,ke)}readViaStream(e){const t=cs(xe(this,ke));let s=Number(e);return new Rc({blockingRead(i){if(s===t.byteLength)throw{tag:"closed"};const r=t.slice(s,s+Number(i));return s+=r.byteLength,r}})}writeViaStream(e){const t=xe(this,ke);let s=Number(e);return new Lc({write(i){const r=new Uint8Array(i.byteLength+t.source.byteLength);return r.set(t.source,0),r.set(i,s),s+=i.byteLength,t.source=r,i.byteLength}})}appendViaStream(){console.log("[filesystem] APPEND STREAM")}advise(e,t,s,i){console.log("[filesystem] ADVISE",e,t,s,i)}syncData(){console.log("[filesystem] SYNC DATA")}getFlags(){console.log("[filesystem] FLAGS FOR")}getType(){return xe(this,qn)?"fifo":xe(this,ke).dir?"directory":xe(this,ke).source?"regular-file":"unknown"}setSize(e){console.log("[filesystem] SET SIZE",e)}setTimes(e,t){console.log("[filesystem] SET TIMES",e,t)}read(e,t){const s=cs(xe(this,ke));return[s.slice(t,t+e),t+e>=s.byteLength]}write(e,t){if(t!==0)throw"invalid-seek";return xe(this,ke).source=e,e.byteLength}readDirectory(){var e;if(!((e=xe(this,ke))!=null&&e.dir))throw"bad-descriptor";return new Da(Object.entries(xe(this,ke).dir).sort(([t],[s])=>t>s?1:-1))}sync(){console.log("[filesystem] SYNC")}createDirectoryAt(e){if(ki(xe(this,ke),e,{create:!0,directory:!0}).source)throw"exist"}stat(){let e="unknown",t=BigInt(0);if(xe(this,ke).source)e="directory";else if(xe(this,ke).dir){e="regular-file";const s=cs(xe(this,ke));t=BigInt(s.byteLength)}return{type:e,linkCount:BigInt(0),size:t,dataAccessTimestamp:Wt,dataModificationTimestamp:Wt,statusChangeTimestamp:Wt}}statAt(e,t){const s=ki(xe(this,ke),t);let i="unknown",r=BigInt(0);if(s.source){i="regular-file";const o=cs(s);r=BigInt(o.byteLength)}else s.dir&&(i="directory");return{type:i,linkCount:BigInt(0),size:r,dataAccessTimestamp:Wt,dataModificationTimestamp:Wt,statusChangeTimestamp:Wt}}setTimesAt(){console.log("[filesystem] SET TIMES AT")}linkAt(){console.log("[filesystem] LINK AT")}openAt(e,t,s,i,r){const o=ki(xe(this,ke),t,s);return new Zn(o)}readlinkAt(){console.log("[filesystem] READLINK AT")}removeDirectoryAt(){console.log("[filesystem] REMOVE DIR AT")}renameAt(){console.log("[filesystem] RENAME AT")}symlinkAt(){console.log("[filesystem] SYMLINK AT")}unlinkFileAt(){console.log("[filesystem] UNLINK FILE AT")}isSameObject(e){return e===this}metadataHash(){let e=BigInt(0);return e+=BigInt(xe(this,Xn)),{upper:e,lower:BigInt(0)}}metadataHashAt(e,t){let s=BigInt(0);return s+=BigInt(xe(this,Xn)),{upper:s,lower:BigInt(0)}}},qn=new WeakMap,ke=new WeakMap,Xn=new WeakMap,Zn);const Tc=Zs.prototype._getEntry;delete Zs.prototype._getEntry;let Ta=[[new Zs(Dc),"/"]],io=Ta[0];const Bc={getDirectories(){return Ta}},_c={Descriptor:Zs,DirectoryEntryStream:Da},{InputStream:Ba,OutputStream:Ys}=Ir,$r=Symbol.dispose??Symbol.for("dispose");let zc=[],Vc=[],Fc=null;const jc={getEnvironment(){return zc},getArguments(){return Vc},initialCwd(){return Fc}};class Uc extends Error{constructor(e){super(`Component exited ${e?"successfully":"with error"}`),this.exitError=!0,this.ok=e}}const Gc={exit(n){throw new Uc(n.tag==="err")}},Wc=new Ba({blockingRead(n){},subscribe(){},[$r](){}});let _a=new TextDecoder;const Hc=new Ys({write(n){console.log(_a.decode(n))},blockingFlush(){},[$r](){}}),qc=new Ys({write(n){console.error(_a.decode(n))},blockingFlush(){},[$r](){}}),Xc={InputStream:Ba,getStdin(){return Wc}},Zc={OutputStream:Ys,getStdout(){return Hc}},Yc={OutputStream:Ys,getStderr(){return qc}};let Sr=class{},Yn=class{};const Qc=new Yn,Kc=new Yn,Jc=new Sr,ed={TerminalInput:Sr},td={TerminalOutput:Yn},nd={TerminalOutput:Yn,getTerminalStderr(){return Kc}},sd={TerminalInput:Sr,getTerminalStdin(){return Jc}},id={TerminalOutput:Yn,getTerminalStdout(){return Qc}},Ii=65536;let $i,ro;const Ki={getRandomBytes(n){const e=new Uint8Array(Number(n));if(n>Ii)for(var t=0;t<n;t+=Ii)crypto.getRandomValues(e.slice(t,t+Ii));else crypto.getRandomValues(e);return e},getRandomU64(){return crypto.getRandomValues(new BigUint64Array(1))[0]},insecureRandom(){return $i===void 0&&($i=Ki.getRandomU64(),ro=Ki.getRandomU64()),[$i,ro]}},{getEnvironment:rd}=jc,{exit:od}=Gc,{getStderr:ad}=Yc,{getStdin:ld}=Xc,{getStdout:cd}=Zc,{TerminalInput:dd}=ed,{TerminalOutput:za}=td,{getTerminalStderr:ud}=nd,{getTerminalStdin:hd}=sd,{getTerminalStdout:pd}=id,{getDirectories:fd}=Bc,{Descriptor:ct,DirectoryEntryStream:Cf,filesystemErrorCode:md}=_c,{Error:un}=Mc,{InputStream:Qs,OutputStream:Nt}=Ir,{getRandomBytes:gd}=Ki,oo=n=>WebAssembly.compile(typeof Buffer<"u"?Buffer.from(n,"base64"):Uint8Array.from(atob(n),e=>e.charCodeAt(0)));class yd extends Error{constructor(e){const t=typeof e!="string";super(t?`${String(e)} (see error.payload)`:e),Object.defineProperty(this,"payload",{value:e,enumerable:t})}}let Si=new DataView(new ArrayBuffer);const k=n=>Si.buffer===n.buffer?Si:Si=new DataView(n.buffer),bd=typeof process<"u"&&process.versions&&process.versions.node;let Ni;async function ao(n){return bd?(Ni=Ni||await Ra(()=>import("./__vite-browser-external-XPXQPciS.js"),__vite__mapDeps([])),WebAssembly.compile(await Ni.readFile(n))):fetch(n).then(WebAssembly.compileStreaming)}function je(n){return n&&Ed.call(n,"payload")?n.payload:n}const Ed=Object.prototype.hasOwnProperty,ds=WebAssembly.instantiate,Me=Symbol.dispose||Symbol.for("dispose");function Ad(){throw new TypeError("Wasm uninitialized use `await $init` first")}const bt=n=>BigInt.asUintN(64,BigInt(n));function xs(n){return n>>>0}const In=new TextDecoder,wd=new TextEncoder;let wt=0;function tn(n,e,t){if(typeof n!="string")throw new TypeError("expected a string");if(n.length===0)return wt=0,1;let s=0,i=0,r=0;for(;n.length>0;){i=e(i,s,1,s+=n.length*2);const{read:o,written:a}=wd.encodeInto(n,new Uint8Array(t.buffer,i+r,s-r));r+=a,n=n.slice(o)}return wt=r,i}let Y,mt;function vd(){const n=ad();if(!(n instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var e=Ks++;return Qe.set(e,{rep:n,own:!0}),e}function xd(n){let e;switch(n){case 0:{e={tag:"ok",val:void 0};break}case 1:{e={tag:"err",val:void 0};break}default:throw new TypeError("invalid variant discriminant for expected")}od(e)}function kd(){const n=ld();if(!(n instanceof Qs))throw new Error('Resource error: Not a valid "InputStream" resource.');var e=Fa++;return ln.set(e,{rep:n,own:!0}),e}function Id(){const n=cd();if(!(n instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var e=Ks++;return Qe.set(e,{rep:n,own:!0}),e}let Be;function $d(n){var t=fd(),s=t.length,i=lt(0,0,4,s*12);for(let u=0;u<t.length;u++){const p=t[u],h=i+u*12;var[r,o]=p;if(!(r instanceof ct))throw new Error('Resource error: Not a valid "Descriptor" resource.');var a=ja++;He.set(a,{rep:r,own:!0}),k(v).setInt32(h+0,a,!0);var l=tn(o,lt,v),c=wt;k(v).setInt32(h+8,c,!0),k(v).setInt32(h+4,l,!0)}k(v).setInt32(n+4,s,!0),k(v).setInt32(n+0,i,!0)}let v,lt;function Sd(n,e,t){var s=n,i=He.get(s).rep;let r;try{r={tag:"ok",val:ct.prototype.readViaStream.call(i,BigInt.asUintN(64,e))}}catch(c){r={tag:"err",val:je(c)}}var o=r;switch(o.tag){case"ok":{const c=o.val;if(k(v).setInt8(t+0,0,!0),!(c instanceof Qs))throw new Error('Resource error: Not a valid "InputStream" resource.');var a=Fa++;ln.set(a,{rep:c,own:!0}),k(v).setInt32(t+4,a,!0);break}case"err":{const c=o.val;k(v).setInt8(t+0,1,!0);var l=c;let u;switch(l){case"access":{u=0;break}case"would-block":{u=1;break}case"already":{u=2;break}case"bad-descriptor":{u=3;break}case"busy":{u=4;break}case"deadlock":{u=5;break}case"quota":{u=6;break}case"exist":{u=7;break}case"file-too-large":{u=8;break}case"illegal-byte-sequence":{u=9;break}case"in-progress":{u=10;break}case"interrupted":{u=11;break}case"invalid":{u=12;break}case"io":{u=13;break}case"is-directory":{u=14;break}case"loop":{u=15;break}case"too-many-links":{u=16;break}case"message-size":{u=17;break}case"name-too-long":{u=18;break}case"no-device":{u=19;break}case"no-entry":{u=20;break}case"no-lock":{u=21;break}case"insufficient-memory":{u=22;break}case"insufficient-space":{u=23;break}case"not-directory":{u=24;break}case"not-empty":{u=25;break}case"not-recoverable":{u=26;break}case"unsupported":{u=27;break}case"no-tty":{u=28;break}case"no-such-device":{u=29;break}case"overflow":{u=30;break}case"not-permitted":{u=31;break}case"pipe":{u=32;break}case"read-only":{u=33;break}case"invalid-seek":{u=34;break}case"text-file-busy":{u=35;break}case"cross-device":{u=36;break}default:throw c instanceof Error&&console.error(c),new TypeError(`"${l}" is not one of the cases of error-code`)}k(v).setInt8(t+4,u,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Nd(n,e,t){var s=n,i=He.get(s).rep;let r;try{r={tag:"ok",val:ct.prototype.writeViaStream.call(i,BigInt.asUintN(64,e))}}catch(c){r={tag:"err",val:je(c)}}var o=r;switch(o.tag){case"ok":{const c=o.val;if(k(v).setInt8(t+0,0,!0),!(c instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var a=Ks++;Qe.set(a,{rep:c,own:!0}),k(v).setInt32(t+4,a,!0);break}case"err":{const c=o.val;k(v).setInt8(t+0,1,!0);var l=c;let u;switch(l){case"access":{u=0;break}case"would-block":{u=1;break}case"already":{u=2;break}case"bad-descriptor":{u=3;break}case"busy":{u=4;break}case"deadlock":{u=5;break}case"quota":{u=6;break}case"exist":{u=7;break}case"file-too-large":{u=8;break}case"illegal-byte-sequence":{u=9;break}case"in-progress":{u=10;break}case"interrupted":{u=11;break}case"invalid":{u=12;break}case"io":{u=13;break}case"is-directory":{u=14;break}case"loop":{u=15;break}case"too-many-links":{u=16;break}case"message-size":{u=17;break}case"name-too-long":{u=18;break}case"no-device":{u=19;break}case"no-entry":{u=20;break}case"no-lock":{u=21;break}case"insufficient-memory":{u=22;break}case"insufficient-space":{u=23;break}case"not-directory":{u=24;break}case"not-empty":{u=25;break}case"not-recoverable":{u=26;break}case"unsupported":{u=27;break}case"no-tty":{u=28;break}case"no-such-device":{u=29;break}case"overflow":{u=30;break}case"not-permitted":{u=31;break}case"pipe":{u=32;break}case"read-only":{u=33;break}case"invalid-seek":{u=34;break}case"text-file-busy":{u=35;break}case"cross-device":{u=36;break}default:throw c instanceof Error&&console.error(c),new TypeError(`"${l}" is not one of the cases of error-code`)}k(v).setInt8(t+4,u,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Pd(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.appendViaStream.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{const l=r.val;if(k(v).setInt8(e+0,0,!0),!(l instanceof Nt))throw new Error('Resource error: Not a valid "OutputStream" resource.');var o=Ks++;Qe.set(o,{rep:l,own:!0}),k(v).setInt32(e+4,o,!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var a=l;let c;switch(a){case"access":{c=0;break}case"would-block":{c=1;break}case"already":{c=2;break}case"bad-descriptor":{c=3;break}case"busy":{c=4;break}case"deadlock":{c=5;break}case"quota":{c=6;break}case"exist":{c=7;break}case"file-too-large":{c=8;break}case"illegal-byte-sequence":{c=9;break}case"in-progress":{c=10;break}case"interrupted":{c=11;break}case"invalid":{c=12;break}case"io":{c=13;break}case"is-directory":{c=14;break}case"loop":{c=15;break}case"too-many-links":{c=16;break}case"message-size":{c=17;break}case"name-too-long":{c=18;break}case"no-device":{c=19;break}case"no-entry":{c=20;break}case"no-lock":{c=21;break}case"insufficient-memory":{c=22;break}case"insufficient-space":{c=23;break}case"not-directory":{c=24;break}case"not-empty":{c=25;break}case"not-recoverable":{c=26;break}case"unsupported":{c=27;break}case"no-tty":{c=28;break}case"no-such-device":{c=29;break}case"overflow":{c=30;break}case"not-permitted":{c=31;break}case"pipe":{c=32;break}case"read-only":{c=33;break}case"invalid-seek":{c=34;break}case"text-file-busy":{c=35;break}case"cross-device":{c=36;break}default:throw l instanceof Error&&console.error(l),new TypeError(`"${a}" is not one of the cases of error-code`)}k(v).setInt8(e+4,c,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Cd(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.getType.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{const l=r.val;k(v).setInt8(e+0,0,!0);var o=l;let c;switch(o){case"unknown":{c=0;break}case"block-device":{c=1;break}case"character-device":{c=2;break}case"directory":{c=3;break}case"fifo":{c=4;break}case"symbolic-link":{c=5;break}case"regular-file":{c=6;break}case"socket":{c=7;break}default:throw l instanceof Error&&console.error(l),new TypeError(`"${o}" is not one of the cases of descriptor-type`)}k(v).setInt8(e+1,c,!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var a=l;let c;switch(a){case"access":{c=0;break}case"would-block":{c=1;break}case"already":{c=2;break}case"bad-descriptor":{c=3;break}case"busy":{c=4;break}case"deadlock":{c=5;break}case"quota":{c=6;break}case"exist":{c=7;break}case"file-too-large":{c=8;break}case"illegal-byte-sequence":{c=9;break}case"in-progress":{c=10;break}case"interrupted":{c=11;break}case"invalid":{c=12;break}case"io":{c=13;break}case"is-directory":{c=14;break}case"loop":{c=15;break}case"too-many-links":{c=16;break}case"message-size":{c=17;break}case"name-too-long":{c=18;break}case"no-device":{c=19;break}case"no-entry":{c=20;break}case"no-lock":{c=21;break}case"insufficient-memory":{c=22;break}case"insufficient-space":{c=23;break}case"not-directory":{c=24;break}case"not-empty":{c=25;break}case"not-recoverable":{c=26;break}case"unsupported":{c=27;break}case"no-tty":{c=28;break}case"no-such-device":{c=29;break}case"overflow":{c=30;break}case"not-permitted":{c=31;break}case"pipe":{c=32;break}case"read-only":{c=33;break}case"invalid-seek":{c=34;break}case"text-file-busy":{c=35;break}case"cross-device":{c=36;break}default:throw l instanceof Error&&console.error(l),new TypeError(`"${a}" is not one of the cases of error-code`)}k(v).setInt8(e+1,c,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Md(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.stat.call(s)}}catch(P){i={tag:"err",val:je(P)}}var r=i;switch(r.tag){case"ok":{const P=r.val;k(v).setInt8(e+0,0,!0);var{type:o,linkCount:a,size:l,dataAccessTimestamp:c,dataModificationTimestamp:u,statusChangeTimestamp:p}=P,h=o;let S;switch(h){case"unknown":{S=0;break}case"block-device":{S=1;break}case"character-device":{S=2;break}case"directory":{S=3;break}case"fifo":{S=4;break}case"symbolic-link":{S=5;break}case"regular-file":{S=6;break}case"socket":{S=7;break}default:throw o instanceof Error&&console.error(o),new TypeError(`"${h}" is not one of the cases of descriptor-type`)}k(v).setInt8(e+8,S,!0),k(v).setBigInt64(e+16,bt(a),!0),k(v).setBigInt64(e+24,bt(l),!0);var f=c;if(f==null)k(v).setInt8(e+32,0,!0);else{const I=f;k(v).setInt8(e+32,1,!0);var{seconds:A,nanoseconds:m}=I;k(v).setBigInt64(e+40,bt(A),!0),k(v).setInt32(e+48,xs(m),!0)}var y=u;if(y==null)k(v).setInt8(e+56,0,!0);else{const I=y;k(v).setInt8(e+56,1,!0);var{seconds:b,nanoseconds:x}=I;k(v).setBigInt64(e+64,bt(b),!0),k(v).setInt32(e+72,xs(x),!0)}var g=p;if(g==null)k(v).setInt8(e+80,0,!0);else{const I=g;k(v).setInt8(e+80,1,!0);var{seconds:E,nanoseconds:w}=I;k(v).setBigInt64(e+88,bt(E),!0),k(v).setInt32(e+96,xs(w),!0)}break}case"err":{const P=r.val;k(v).setInt8(e+0,1,!0);var $=P;let S;switch($){case"access":{S=0;break}case"would-block":{S=1;break}case"already":{S=2;break}case"bad-descriptor":{S=3;break}case"busy":{S=4;break}case"deadlock":{S=5;break}case"quota":{S=6;break}case"exist":{S=7;break}case"file-too-large":{S=8;break}case"illegal-byte-sequence":{S=9;break}case"in-progress":{S=10;break}case"interrupted":{S=11;break}case"invalid":{S=12;break}case"io":{S=13;break}case"is-directory":{S=14;break}case"loop":{S=15;break}case"too-many-links":{S=16;break}case"message-size":{S=17;break}case"name-too-long":{S=18;break}case"no-device":{S=19;break}case"no-entry":{S=20;break}case"no-lock":{S=21;break}case"insufficient-memory":{S=22;break}case"insufficient-space":{S=23;break}case"not-directory":{S=24;break}case"not-empty":{S=25;break}case"not-recoverable":{S=26;break}case"unsupported":{S=27;break}case"no-tty":{S=28;break}case"no-such-device":{S=29;break}case"overflow":{S=30;break}case"not-permitted":{S=31;break}case"pipe":{S=32;break}case"read-only":{S=33;break}case"invalid-seek":{S=34;break}case"text-file-busy":{S=35;break}case"cross-device":{S=36;break}default:throw P instanceof Error&&console.error(P),new TypeError(`"${$}" is not one of the cases of error-code`)}k(v).setInt8(e+8,S,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Rd(n,e,t,s,i,r,o){var a=n,l=He.get(a).rep;if(e&4294967294)throw new TypeError("flags have extraneous bits set");var c={symlinkFollow:!!(e&1)},u=t,p=s,h=In.decode(new Uint8Array(v.buffer,u,p));if(i&4294967280)throw new TypeError("flags have extraneous bits set");var f={create:!!(i&1),directory:!!(i&2),exclusive:!!(i&4),truncate:!!(i&8)};if(r&4294967232)throw new TypeError("flags have extraneous bits set");var A={read:!!(r&1),write:!!(r&2),fileIntegritySync:!!(r&4),dataIntegritySync:!!(r&8),requestedWriteSync:!!(r&16),mutateDirectory:!!(r&32)};let m;try{m={tag:"ok",val:ct.prototype.openAt.call(l,c,h,f,A)}}catch(g){m={tag:"err",val:je(g)}}var y=m;switch(y.tag){case"ok":{const g=y.val;if(k(v).setInt8(o+0,0,!0),!(g instanceof ct))throw new Error('Resource error: Not a valid "Descriptor" resource.');var b=ja++;He.set(b,{rep:g,own:!0}),k(v).setInt32(o+4,b,!0);break}case"err":{const g=y.val;k(v).setInt8(o+0,1,!0);var x=g;let E;switch(x){case"access":{E=0;break}case"would-block":{E=1;break}case"already":{E=2;break}case"bad-descriptor":{E=3;break}case"busy":{E=4;break}case"deadlock":{E=5;break}case"quota":{E=6;break}case"exist":{E=7;break}case"file-too-large":{E=8;break}case"illegal-byte-sequence":{E=9;break}case"in-progress":{E=10;break}case"interrupted":{E=11;break}case"invalid":{E=12;break}case"io":{E=13;break}case"is-directory":{E=14;break}case"loop":{E=15;break}case"too-many-links":{E=16;break}case"message-size":{E=17;break}case"name-too-long":{E=18;break}case"no-device":{E=19;break}case"no-entry":{E=20;break}case"no-lock":{E=21;break}case"insufficient-memory":{E=22;break}case"insufficient-space":{E=23;break}case"not-directory":{E=24;break}case"not-empty":{E=25;break}case"not-recoverable":{E=26;break}case"unsupported":{E=27;break}case"no-tty":{E=28;break}case"no-such-device":{E=29;break}case"overflow":{E=30;break}case"not-permitted":{E=31;break}case"pipe":{E=32;break}case"read-only":{E=33;break}case"invalid-seek":{E=34;break}case"text-file-busy":{E=35;break}case"cross-device":{E=36;break}default:throw g instanceof Error&&console.error(g),new TypeError(`"${x}" is not one of the cases of error-code`)}k(v).setInt8(o+4,E,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Ld(n,e){var t=n,s=He.get(t).rep;let i;try{i={tag:"ok",val:ct.prototype.metadataHash.call(s)}}catch(c){i={tag:"err",val:je(c)}}var r=i;switch(r.tag){case"ok":{const c=r.val;k(v).setInt8(e+0,0,!0);var{lower:o,upper:a}=c;k(v).setBigInt64(e+8,bt(o),!0),k(v).setBigInt64(e+16,bt(a),!0);break}case"err":{const c=r.val;k(v).setInt8(e+0,1,!0);var l=c;let u;switch(l){case"access":{u=0;break}case"would-block":{u=1;break}case"already":{u=2;break}case"bad-descriptor":{u=3;break}case"busy":{u=4;break}case"deadlock":{u=5;break}case"quota":{u=6;break}case"exist":{u=7;break}case"file-too-large":{u=8;break}case"illegal-byte-sequence":{u=9;break}case"in-progress":{u=10;break}case"interrupted":{u=11;break}case"invalid":{u=12;break}case"io":{u=13;break}case"is-directory":{u=14;break}case"loop":{u=15;break}case"too-many-links":{u=16;break}case"message-size":{u=17;break}case"name-too-long":{u=18;break}case"no-device":{u=19;break}case"no-entry":{u=20;break}case"no-lock":{u=21;break}case"insufficient-memory":{u=22;break}case"insufficient-space":{u=23;break}case"not-directory":{u=24;break}case"not-empty":{u=25;break}case"not-recoverable":{u=26;break}case"unsupported":{u=27;break}case"no-tty":{u=28;break}case"no-such-device":{u=29;break}case"overflow":{u=30;break}case"not-permitted":{u=31;break}case"pipe":{u=32;break}case"read-only":{u=33;break}case"invalid-seek":{u=34;break}case"text-file-busy":{u=35;break}case"cross-device":{u=36;break}default:throw c instanceof Error&&console.error(c),new TypeError(`"${l}" is not one of the cases of error-code`)}k(v).setInt8(e+8,u,!0);break}default:throw new TypeError("invalid variant specified for result")}}function Od(n,e){var t=n,s=dt.get(t).rep,r=md(s);if(r==null)k(v).setInt8(e+0,0,!0);else{const a=r;k(v).setInt8(e+0,1,!0);var o=a;let l;switch(o){case"access":{l=0;break}case"would-block":{l=1;break}case"already":{l=2;break}case"bad-descriptor":{l=3;break}case"busy":{l=4;break}case"deadlock":{l=5;break}case"quota":{l=6;break}case"exist":{l=7;break}case"file-too-large":{l=8;break}case"illegal-byte-sequence":{l=9;break}case"in-progress":{l=10;break}case"interrupted":{l=11;break}case"invalid":{l=12;break}case"io":{l=13;break}case"is-directory":{l=14;break}case"loop":{l=15;break}case"too-many-links":{l=16;break}case"message-size":{l=17;break}case"name-too-long":{l=18;break}case"no-device":{l=19;break}case"no-entry":{l=20;break}case"no-lock":{l=21;break}case"insufficient-memory":{l=22;break}case"insufficient-space":{l=23;break}case"not-directory":{l=24;break}case"not-empty":{l=25;break}case"not-recoverable":{l=26;break}case"unsupported":{l=27;break}case"no-tty":{l=28;break}case"no-such-device":{l=29;break}case"overflow":{l=30;break}case"not-permitted":{l=31;break}case"pipe":{l=32;break}case"read-only":{l=33;break}case"invalid-seek":{l=34;break}case"text-file-busy":{l=35;break}case"cross-device":{l=36;break}default:throw a instanceof Error&&console.error(a),new TypeError(`"${o}" is not one of the cases of error-code`)}k(v).setInt8(e+1,l,!0)}}function Dd(n,e,t){var s=n,i=ln.get(s).rep;let r;try{r={tag:"ok",val:Qs.prototype.read.call(i,BigInt.asUintN(64,e))}}catch(f){r={tag:"err",val:je(f)}}var o=r;switch(o.tag){case"ok":{const f=o.val;k(v).setInt8(t+0,0,!0);var a=f,l=a.byteLength,c=lt(0,0,1,l*1),u=new Uint8Array(a.buffer||a,a.byteOffset,l*1);new Uint8Array(v.buffer,c,l*1).set(u),k(v).setInt32(t+8,l,!0),k(v).setInt32(t+4,c,!0);break}case"err":{const f=o.val;k(v).setInt8(t+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(t+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(t+8,h,!0);break}case"closed":{k(v).setInt8(t+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Td(n,e,t){var s=n,i=ln.get(s).rep;let r;try{r={tag:"ok",val:Qs.prototype.blockingRead.call(i,BigInt.asUintN(64,e))}}catch(f){r={tag:"err",val:je(f)}}var o=r;switch(o.tag){case"ok":{const f=o.val;k(v).setInt8(t+0,0,!0);var a=f,l=a.byteLength,c=lt(0,0,1,l*1),u=new Uint8Array(a.buffer||a,a.byteOffset,l*1);new Uint8Array(v.buffer,c,l*1).set(u),k(v).setInt32(t+8,l,!0),k(v).setInt32(t+4,c,!0);break}case"err":{const f=o.val;k(v).setInt8(t+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(t+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(t+8,h,!0);break}case"closed":{k(v).setInt8(t+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Bd(n,e){var t=n,s=Qe.get(t).rep;let i;try{i={tag:"ok",val:Nt.prototype.checkWrite.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{const l=r.val;k(v).setInt8(e+0,0,!0),k(v).setBigInt64(e+8,bt(l),!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var o=l;switch(o.tag){case"last-operation-failed":{const c=o.val;if(k(v).setInt8(e+8,0,!0),!(c instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var a=hn++;dt.set(a,{rep:c,own:!0}),k(v).setInt32(e+12,a,!0);break}case"closed":{k(v).setInt8(e+8,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(o.tag)}\` (received \`${o}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function _d(n,e,t,s){var i=n,r=Qe.get(i).rep,o=e,a=t,l=new Uint8Array(v.buffer.slice(o,o+a*1));let c;try{c={tag:"ok",val:Nt.prototype.write.call(r,l)}}catch(f){c={tag:"err",val:je(f)}}var u=c;switch(u.tag){case"ok":{u.val,k(v).setInt8(s+0,0,!0);break}case"err":{const f=u.val;k(v).setInt8(s+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(s+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(s+8,h,!0);break}case"closed":{k(v).setInt8(s+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function zd(n,e,t,s){var i=n,r=Qe.get(i).rep,o=e,a=t,l=new Uint8Array(v.buffer.slice(o,o+a*1));let c;try{c={tag:"ok",val:Nt.prototype.blockingWriteAndFlush.call(r,l)}}catch(f){c={tag:"err",val:je(f)}}var u=c;switch(u.tag){case"ok":{u.val,k(v).setInt8(s+0,0,!0);break}case"err":{const f=u.val;k(v).setInt8(s+0,1,!0);var p=f;switch(p.tag){case"last-operation-failed":{const A=p.val;if(k(v).setInt8(s+4,0,!0),!(A instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var h=hn++;dt.set(h,{rep:A,own:!0}),k(v).setInt32(s+8,h,!0);break}case"closed":{k(v).setInt8(s+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(p.tag)}\` (received \`${p}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Vd(n,e){var t=n,s=Qe.get(t).rep;let i;try{i={tag:"ok",val:Nt.prototype.blockingFlush.call(s)}}catch(l){i={tag:"err",val:je(l)}}var r=i;switch(r.tag){case"ok":{r.val,k(v).setInt8(e+0,0,!0);break}case"err":{const l=r.val;k(v).setInt8(e+0,1,!0);var o=l;switch(o.tag){case"last-operation-failed":{const c=o.val;if(k(v).setInt8(e+4,0,!0),!(c instanceof un))throw new Error('Resource error: Not a valid "Error" resource.');var a=hn++;dt.set(a,{rep:c,own:!0}),k(v).setInt32(e+8,a,!0);break}case"closed":{k(v).setInt8(e+4,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify(o.tag)}\` (received \`${o}\`) specified for \`StreamError\``)}break}default:throw new TypeError("invalid variant specified for result")}}function Fd(n,e){var s=gd(BigInt.asUintN(64,n)),i=s.byteLength,r=lt(0,0,1,i*1),o=new Uint8Array(s.buffer||s,s.byteOffset,i*1);new Uint8Array(v.buffer,r,i*1).set(o),k(v).setInt32(e+4,i,!0),k(v).setInt32(e+0,r,!0)}function jd(n){var t=rd(),s=t.length,i=lt(0,0,4,s*16);for(let p=0;p<t.length;p++){const h=t[p],f=i+p*16;var[r,o]=h,a=tn(r,lt,v),l=wt;k(v).setInt32(f+4,l,!0),k(v).setInt32(f+0,a,!0);var c=tn(o,lt,v),u=wt;k(v).setInt32(f+12,u,!0),k(v).setInt32(f+8,c,!0)}k(v).setInt32(n+4,s,!0),k(v).setInt32(n+0,i,!0)}function Ud(n){var t=hd();if(t==null)k(v).setInt8(n+0,0,!0);else{const i=t;if(k(v).setInt8(n+0,1,!0),!(i instanceof dd))throw new Error('Resource error: Not a valid "TerminalInput" resource.');var s=tu++;Ji.set(s,{rep:i,own:!0}),k(v).setInt32(n+4,s,!0)}}function Gd(n){var t=pd();if(t==null)k(v).setInt8(n+0,0,!0);else{const i=t;if(k(v).setInt8(n+0,1,!0),!(i instanceof za))throw new Error('Resource error: Not a valid "TerminalOutput" resource.');var s=Ua++;Rs.set(s,{rep:i,own:!0}),k(v).setInt32(n+4,s,!0)}}function Wd(n){var t=ud();if(t==null)k(v).setInt8(n+0,0,!0);else{const i=t;if(k(v).setInt8(n+0,1,!0),!(i instanceof za))throw new Error('Resource error: Not a valid "TerminalOutput" resource.');var s=Ua++;Rs.set(s,{rep:i,own:!0}),k(v).setInt32(n+4,s,!0)}}let Ct,Va;function Hd(n){const e=lo.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);lo.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function qd(n){const e=dt.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);dt.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Xd(n){const e=ln.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);ln.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Zd(n){const e=Qe.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);Qe.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Yd(n){const e=He.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);He.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Qd(n){const e=co.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);co.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Kd(n){const e=Rs.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);Rs.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function Jd(n){const e=Ji.get(n);if(!e)throw new Error(`Resource error: Invalid handle ${n}`);Ji.delete(n),e.own&&e.rep[Me]&&e.rep[Me]()}function eu(n,e){Ga||Ad();var t=Ct(0,0,4,52),s=n,i=s.byteLength,r=Ct(0,0,1,i*1),o=new Uint8Array(s.buffer||s,s.byteOffset,i*1);new Uint8Array(v.buffer,r,i*1).set(o),k(v).setInt32(t+4,i,!0),k(v).setInt32(t+0,r,!0);var{name:a,noTypescript:l,instantiation:c,map:u,compat:p,noNodejsCompat:h,base64Cutoff:f,tlaCompat:A,validLiftingOptimization:m,tracing:y,noNamespacedExports:b}=e,x=tn(a,Ct,v),g=wt;k(v).setInt32(t+12,g,!0),k(v).setInt32(t+8,x,!0);var E=l;if(E==null)k(v).setInt8(t+16,0,!0);else{const Q=E;k(v).setInt8(t+16,1,!0),k(v).setInt8(t+17,Q?1:0,!0)}var w=c;if(w==null)k(v).setInt8(t+18,0,!0);else{const Q=w;k(v).setInt8(t+18,1,!0);var $=Q;switch($.tag){case"async":{k(v).setInt8(t+19,0,!0);break}case"sync":{k(v).setInt8(t+19,1,!0);break}default:throw new TypeError(`invalid variant tag value \`${JSON.stringify($.tag)}\` (received \`${$}\`) specified for \`InstantiationMode\``)}}var P=u;if(P==null)k(v).setInt8(t+20,0,!0);else{const Q=P;k(v).setInt8(t+20,1,!0);var S=Q,I=S.length,C=Ct(0,0,4,I*16);for(let Ce=0;Ce<S.length;Ce++){const An=S[Ce],os=C+Ce*16;var[V,R]=An,_=tn(V,Ct,v),D=wt;k(v).setInt32(os+4,D,!0),k(v).setInt32(os+0,_,!0);var L=tn(R,Ct,v),B=wt;k(v).setInt32(os+12,B,!0),k(v).setInt32(os+8,L,!0)}k(v).setInt32(t+28,I,!0),k(v).setInt32(t+24,C,!0)}var F=p;if(F==null)k(v).setInt8(t+32,0,!0);else{const Q=F;k(v).setInt8(t+32,1,!0),k(v).setInt8(t+33,Q?1:0,!0)}var H=h;if(H==null)k(v).setInt8(t+34,0,!0);else{const Q=H;k(v).setInt8(t+34,1,!0),k(v).setInt8(t+35,Q?1:0,!0)}var oe=f;if(oe==null)k(v).setInt8(t+36,0,!0);else{const Q=oe;k(v).setInt8(t+36,1,!0),k(v).setInt32(t+40,xs(Q),!0)}var K=A;if(K==null)k(v).setInt8(t+44,0,!0);else{const Q=K;k(v).setInt8(t+44,1,!0),k(v).setInt8(t+45,Q?1:0,!0)}var G=m;if(G==null)k(v).setInt8(t+46,0,!0);else{const Q=G;k(v).setInt8(t+46,1,!0),k(v).setInt8(t+47,Q?1:0,!0)}var ae=y;if(ae==null)k(v).setInt8(t+48,0,!0);else{const Q=ae;k(v).setInt8(t+48,1,!0),k(v).setInt8(t+49,Q?1:0,!0)}var Ne=b;if(Ne==null)k(v).setInt8(t+50,0,!0);else{const Q=Ne;k(v).setInt8(t+50,1,!0),k(v).setInt8(t+51,Q?1:0,!0)}const we=mt.generate(t);let he;switch(k(v).getUint8(we+0,!0)){case 0:{var Pe=k(v).getInt32(we+8,!0),tt=k(v).getInt32(we+4,!0),ve=[];for(let Q=0;Q<Pe;Q++){const Ce=tt+Q*16;var ge=k(v).getInt32(Ce+0,!0),pt=k(v).getInt32(Ce+4,!0),yn=In.decode(new Uint8Array(v.buffer,ge,pt)),is=k(v).getInt32(Ce+8,!0),pi=k(v).getInt32(Ce+12,!0),fi=new Uint8Array(v.buffer.slice(is,is+pi*1));ve.push([yn,fi])}var mi=k(v).getInt32(we+16,!0),gi=k(v).getInt32(we+12,!0),rs=[];for(let Q=0;Q<mi;Q++){const Ce=gi+Q*8;var Gt=k(v).getInt32(Ce+0,!0),Re=k(v).getInt32(Ce+4,!0),bn=In.decode(new Uint8Array(v.buffer,Gt,Re));rs.push(bn)}var yi=k(v).getInt32(we+24,!0),bi=k(v).getInt32(we+20,!0),En=[];for(let Q=0;Q<yi;Q++){const Ce=bi+Q*12;var Ei=k(v).getInt32(Ce+0,!0),Ai=k(v).getInt32(Ce+4,!0),wi=In.decode(new Uint8Array(v.buffer,Ei,Ai));let An;switch(k(v).getUint8(Ce+8,!0)){case 0:{An="function";break}case 1:{An="instance";break}default:throw new TypeError("invalid discriminant specified for ExportType")}En.push([wi,An])}he={tag:"ok",val:{files:ve,imports:rs,exports:En}};break}case 1:{var vc=k(v).getInt32(we+4,!0),xc=k(v).getInt32(we+8,!0),kc=In.decode(new Uint8Array(v.buffer,vc,xc));he={tag:"err",val:kc};break}default:throw new TypeError("invalid variant discriminant for expected")}if(Va(we),he.tag==="err")throw new yd(he.val);return he.val}const dt=new Map;let hn=0;const ln=new Map;let Fa=0;const Qe=new Map;let Ks=0;const He=new Map;let ja=0;const lo=new Map,co=new Map,Ji=new Map;let tu=0;const Rs=new Map;let Ua=0,Ga=!1;const nu=(async()=>{const n=ao(new URL("/loader/assets/js-component-bindgen-component.core-8p6hfTFd.wasm",import.meta.url)),e=ao(new URL("/loader/assets/js-component-bindgen-component.core2-0P6l7xr-.wasm",import.meta.url)),t=oo("AGFzbQEAAAABUQxgAX8AYAN/fn8AYAJ/fwBgB39/f39/f38AYAR/f39/AGACfn8AYAl/f39/f35+f38Bf2ACf38Bf2AEf39/fwF/YAF/AX9gA39/fwF/YAF/AAMgHwABAQICAgMCAgEBAgQEAgUAAAAABgcHCAgHBwkHCgsEBQFwAR8fB50BIAEwAAABMQABATIAAgEzAAMBNAAEATUABQE2AAYBNwAHATgACAE5AAkCMTAACgIxMQALAjEyAAwCMTMADQIxNAAOAjE1AA8CMTYAEAIxNwARAjE4ABICMTkAEwIyMAAUAjIxABUCMjIAFgIyMwAXAjI0ABgCMjUAGQIyNgAaAjI3ABsCMjgAHAIyOQAdAjMwAB4IJGltcG9ydHMBAAqZAx8JACAAQQARAAALDQAgACABIAJBAREBAAsNACAAIAEgAkECEQEACwsAIAAgAUEDEQIACwsAIAAgAUEEEQIACwsAIAAgAUEFEQIACxUAIAAgASACIAMgBCAFIAZBBhEDAAsLACAAIAFBBxECAAsLACAAIAFBCBECAAsNACAAIAEgAkEJEQEACw0AIAAgASACQQoRAQALCwAgACABQQsRAgALDwAgACABIAIgA0EMEQQACw8AIAAgASACIANBDREEAAsLACAAIAFBDhECAAsLACAAIAFBDxEFAAsJACAAQRARAAALCQAgAEEREQAACwkAIABBEhEAAAsJACAAQRMRAAALGQAgACABIAIgAyAEIAUgBiAHIAhBFBEGAAsLACAAIAFBFREHAAsLACAAIAFBFhEHAAsPACAAIAEgAiADQRcRCAALDwAgACABIAIgA0EYEQgACwsAIAAgAUEZEQcACwsAIAAgAUEaEQcACwkAIABBGxEJAAsLACAAIAFBHBEHAAsNACAAIAEgAkEdEQoACwkAIABBHhELAAsALglwcm9kdWNlcnMBDHByb2Nlc3NlZC1ieQENd2l0LWNvbXBvbmVudAYwLjE5LjAAkhAEbmFtZQATEndpdC1jb21wb25lbnQ6c2hpbQH1Dx8ARWluZGlyZWN0LXdhc2k6ZmlsZXN5c3RlbS9wcmVvcGVuc0AwLjIuMC1yYy0yMDIzLTExLTEwLWdldC1kaXJlY3RvcmllcwFVaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLnJlYWQtdmlhLXN0cmVhbQJWaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLndyaXRlLXZpYS1zdHJlYW0DV2luZGlyZWN0LXdhc2k6ZmlsZXN5c3RlbS90eXBlc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdZGVzY3JpcHRvci5hcHBlbmQtdmlhLXN0cmVhbQROaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLmdldC10eXBlBUppbmRpcmVjdC13YXNpOmZpbGVzeXN0ZW0vdHlwZXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1bbWV0aG9kXWRlc2NyaXB0b3Iuc3RhdAZNaW5kaXJlY3Qtd2FzaTpmaWxlc3lzdGVtL3R5cGVzQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1kZXNjcmlwdG9yLm9wZW4tYXQHU2luZGlyZWN0LXdhc2k6ZmlsZXN5c3RlbS90eXBlc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdZGVzY3JpcHRvci5tZXRhZGF0YS1oYXNoCEhpbmRpcmVjdC13YXNpOmZpbGVzeXN0ZW0vdHlwZXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1maWxlc3lzdGVtLWVycm9yLWNvZGUJRmluZGlyZWN0LXdhc2k6aW8vc3RyZWFtc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdaW5wdXQtc3RyZWFtLnJlYWQKT2luZGlyZWN0LXdhc2k6aW8vc3RyZWFtc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2RdaW5wdXQtc3RyZWFtLmJsb2NraW5nLXJlYWQLTmluZGlyZWN0LXdhc2k6aW8vc3RyZWFtc0AwLjIuMC1yYy0yMDIzLTExLTEwLVttZXRob2Rdb3V0cHV0LXN0cmVhbS5jaGVjay13cml0ZQxIaW5kaXJlY3Qtd2FzaTppby9zdHJlYW1zQDAuMi4wLXJjLTIwMjMtMTEtMTAtW21ldGhvZF1vdXRwdXQtc3RyZWFtLndyaXRlDVtpbmRpcmVjdC13YXNpOmlvL3N0cmVhbXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1bbWV0aG9kXW91dHB1dC1zdHJlYW0uYmxvY2tpbmctd3JpdGUtYW5kLWZsdXNoDlFpbmRpcmVjdC13YXNpOmlvL3N0cmVhbXNAMC4yLjAtcmMtMjAyMy0xMS0xMC1bbWV0aG9kXW91dHB1dC1zdHJlYW0uYmxvY2tpbmctZmx1c2gPQGluZGlyZWN0LXdhc2k6cmFuZG9tL3JhbmRvbUAwLjIuMC1yYy0yMDIzLTExLTEwLWdldC1yYW5kb20tYnl0ZXMQQWluZGlyZWN0LXdhc2k6Y2xpL2Vudmlyb25tZW50QDAuMi4wLXJjLTIwMjMtMTEtMTAtZ2V0LWVudmlyb25tZW50EUdpbmRpcmVjdC13YXNpOmNsaS90ZXJtaW5hbC1zdGRpbkAwLjIuMC1yYy0yMDIzLTExLTEwLWdldC10ZXJtaW5hbC1zdGRpbhJJaW5kaXJlY3Qtd2FzaTpjbGkvdGVybWluYWwtc3Rkb3V0QDAuMi4wLXJjLTIwMjMtMTEtMTAtZ2V0LXRlcm1pbmFsLXN0ZG91dBNJaW5kaXJlY3Qtd2FzaTpjbGkvdGVybWluYWwtc3RkZXJyQDAuMi4wLXJjLTIwMjMtMTEtMTAtZ2V0LXRlcm1pbmFsLXN0ZGVychQmYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1wYXRoX29wZW4VLGFkYXB0LXdhc2lfc25hcHNob3RfcHJldmlldzEtZmRfZmlsZXN0YXRfZ2V0FidhZGFwdC13YXNpX3NuYXBzaG90X3ByZXZpZXcxLXJhbmRvbV9nZXQXJGFkYXB0LXdhc2lfc25hcHNob3RfcHJldmlldzEtZmRfcmVhZBglYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF93cml0ZRkoYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1lbnZpcm9uX2dldBouYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1lbnZpcm9uX3NpemVzX2dldBslYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF9jbG9zZRwrYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF9wcmVzdGF0X2dldB0wYWRhcHQtd2FzaV9zbmFwc2hvdF9wcmV2aWV3MS1mZF9wcmVzdGF0X2Rpcl9uYW1lHiZhZGFwdC13YXNpX3NuYXBzaG90X3ByZXZpZXcxLXByb2NfZXhpdA"),s=oo("AGFzbQEAAAABUQxgAX8AYAN/fn8AYAJ/fwBgB39/f39/f38AYAR/f39/AGACfn8AYAl/f39/f35+f38Bf2ACf38Bf2AEf39/fwF/YAF/AX9gA39/fwF/YAF/AALAASAAATAAAAABMQABAAEyAAEAATMAAgABNAACAAE1AAIAATYAAwABNwACAAE4AAIAATkAAQACMTAAAQACMTEAAgACMTIABAACMTMABAACMTQAAgACMTUABQACMTYAAAACMTcAAAACMTgAAAACMTkAAAACMjAABgACMjEABwACMjIABwACMjMACAACMjQACAACMjUABwACMjYABwACMjcACQACMjgABwACMjkACgACMzAACwAIJGltcG9ydHMBcAEfHwklAQBBAAsfAAECAwQFBgcICQoLDA0ODxAREhMUFRYXGBkaGxwdHgAuCXByb2R1Y2VycwEMcHJvY2Vzc2VkLWJ5AQ13aXQtY29tcG9uZW50BjAuMTkuMAAcBG5hbWUAFRR3aXQtY29tcG9uZW50OmZpeHVwcw");({exports:Y}=await ds(await t)),{exports:mt}=await ds(await n,{wasi_snapshot_preview1:{environ_get:Y[25],environ_sizes_get:Y[26],fd_close:Y[27],fd_filestat_get:Y[21],fd_prestat_dir_name:Y[29],fd_prestat_get:Y[28],fd_read:Y[23],fd_write:Y[24],path_open:Y[20],proc_exit:Y[30],random_get:Y[22]}}),{exports:Be}=await ds(await e,{__main_module__:{cabi_realloc:mt.cabi_realloc},env:{memory:mt.memory},"wasi:cli/environment@0.2.0-rc-2023-11-10":{"get-environment":Y[16]},"wasi:cli/exit@0.2.0-rc-2023-11-10":{exit:xd},"wasi:cli/stderr@0.2.0-rc-2023-11-10":{"get-stderr":vd},"wasi:cli/stdin@0.2.0-rc-2023-11-10":{"get-stdin":kd},"wasi:cli/stdout@0.2.0-rc-2023-11-10":{"get-stdout":Id},"wasi:cli/terminal-input@0.2.0-rc-2023-11-10":{"[resource-drop]terminal-input":Jd},"wasi:cli/terminal-output@0.2.0-rc-2023-11-10":{"[resource-drop]terminal-output":Kd},"wasi:cli/terminal-stderr@0.2.0-rc-2023-11-10":{"get-terminal-stderr":Y[19]},"wasi:cli/terminal-stdin@0.2.0-rc-2023-11-10":{"get-terminal-stdin":Y[17]},"wasi:cli/terminal-stdout@0.2.0-rc-2023-11-10":{"get-terminal-stdout":Y[18]},"wasi:filesystem/preopens@0.2.0-rc-2023-11-10":{"get-directories":Y[0]},"wasi:filesystem/types@0.2.0-rc-2023-11-10":{"[method]descriptor.append-via-stream":Y[3],"[method]descriptor.get-type":Y[4],"[method]descriptor.metadata-hash":Y[7],"[method]descriptor.open-at":Y[6],"[method]descriptor.read-via-stream":Y[1],"[method]descriptor.stat":Y[5],"[method]descriptor.write-via-stream":Y[2],"[resource-drop]descriptor":Yd,"[resource-drop]directory-entry-stream":Hd,"filesystem-error-code":Y[8]},"wasi:io/error@0.2.0-rc-2023-11-10":{"[resource-drop]error":qd},"wasi:io/streams@0.2.0-rc-2023-11-10":{"[method]input-stream.blocking-read":Y[10],"[method]input-stream.read":Y[9],"[method]output-stream.blocking-flush":Y[14],"[method]output-stream.blocking-write-and-flush":Y[13],"[method]output-stream.check-write":Y[11],"[method]output-stream.write":Y[12],"[resource-drop]input-stream":Xd,"[resource-drop]output-stream":Zd},"wasi:random/random@0.2.0-rc-2023-11-10":{"get-random-bytes":Y[15]},"wasi:sockets/tcp@0.2.0-rc-2023-11-10":{"[resource-drop]tcp-socket":Qd}}),v=mt.memory,lt=Be.cabi_import_realloc,await ds(await s,{"":{$imports:Y.$imports,0:$d,1:Sd,10:Td,11:Bd,12:_d,13:zd,14:Vd,15:Fd,16:jd,17:Ud,18:Gd,19:Wd,2:Nd,20:Be.path_open,21:Be.fd_filestat_get,22:Be.random_get,23:Be.fd_read,24:Be.fd_write,25:Be.environ_get,26:Be.environ_sizes_get,27:Be.fd_close,28:Be.fd_prestat_get,29:Be.fd_prestat_dir_name,3:Pd,30:Be.proc_exit,4:Cd,5:Md,6:Rd,7:Ld,8:Od,9:Dd}}),Ct=mt.cabi_realloc,Va=mt.cabi_post_generate,mt["cabi_post_generate-types"],Ga=!0})();async function su(){return await nu,eu.apply(this,arguments)}/*
   @license
 	Rollup.js v4.9.1
 	Sun, 17 Dec 2023 06:25:43 GMT - commit d56ac63dc0452820272a0d7536340277f7db68bf
