+++ services/system/AccountSys/ui/public/loader/index.html
@@ -5,7 +5,7 @@
     <link rel="icon" type="image/svg+xml" href="/loader/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + TS</title>
-    <script type="module" crossorigin src="/loader/assets/index-pZAdW6pa.js"></script>
+    <script type="module" crossorigin src="/loader/assets/index-IMu5n1Ph.js"></script>
   </head>
   <body>
     <div id="app"></div>
