+++ services/user/Loader/src/importables.js
@@ -1,0 +1,7 @@
+function derp(num) {
+  return num * 2;
+}
+
+export function prnt(string) {
+  console.log(`from imported code: john ${derp(2)}`, string);
+}
